refactor API to return bad HTTP status codes
