+++ Client/src/app/app.component.html
@@ -67,7 +67,7 @@
         <mat-chip-list *ngIf="!pickIndicator">
 
             <mat-chip selected color="default">
-                <mat-icon mat-list-icon [style.color]="'black'">show_chart</mat-icon>&nbsp;
+                <mat-icon mat-list-icon [style.color]="'darkgreen'">waterfall_chart</mat-icon>&nbsp;
                 <span mat-line>Price</span>
             </mat-chip>
