+++ Server/WebApi/Controllers/MainController.cs
@@ -1,73 +1,122 @@
-﻿using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Mvc;
 using Skender.Stock.Indicators;
-using System;
-using System.Collections.Generic;
-using System.Linq;
 using WebApi.Services;
 
-namespace WebApi.Controllers
+namespace WebApi.Controllers;
+
+
+[ApiController]
+[Route("")]
+public class MainController : ControllerBase
 {
+    internal static readonly IEnumerable<Quote> quotes = FetchQuotes.Get();
+    internal static readonly DateTime dateStart = DateTime.Parse("6/1/2018");
 
-    [ApiController]
-    [Route("")]
-    public class MainController : ControllerBase
+    [HttpGet]
+    public string Get()
     {
-        internal static readonly IEnumerable<Quote> quotes = FetchQuotes.Get();
-        internal static readonly DateTime dateStart = DateTime.Parse("6/1/2018");
+        return "API is functioning nominally.";
+    }
 
-        [HttpGet]
-        public string Get()
+    [HttpGet("history")]
+    public IEnumerable<Quote> GetQuotes()
+    {
+        return quotes
+            .Where(x => x.Date >= dateStart);
+    }
+
+
+    //////////////////////////////////////////
+    // INDICATORS (sorted alphabetically)
+
+    [HttpGet("BB")]
+    public IActionResult GetBollingerBands(
+         int lookbackPeriods,
+         double standardDeviations)
+    {
+        try
         {
-            return "API is functioning nominally.";
+            IEnumerable<BollingerBandsResult> results =
+                quotes.GetBollingerBands(lookbackPeriods, standardDeviations)
+                      .Where(x => x.Date >= dateStart);
+
+            return Ok(results);
         }
-
-        [HttpGet("history")]
-        public IEnumerable<Quote> GetQuotes()
+        catch (ArgumentOutOfRangeException rex)
         {
-            return quotes
-                .Where(x => x.Date >= dateStart);
+            return BadRequest(rex.Message);
         }
+    }
 
-
-        //////////////////////////////////////////
-        // INDICATORS (sorted alphabetically)
-
-        [HttpGet("BB")]
-        public IEnumerable<BollingerBandsResult> GetBollingerBands(
-             int lookbackPeriods, double standardDeviations)
+    [HttpGet("EMA")]
+    public IActionResult GetEMA(int lookbackPeriods)
+    {
+        try
         {
-            return quotes.GetBollingerBands(lookbackPeriods, standardDeviations)
-                .Where(x => x.Date >= dateStart);
+            IEnumerable<EmaResult> results =
+                quotes.GetEma(lookbackPeriods)
+                      .Where(x => x.Date >= dateStart);
+
+            return Ok(results);
         }
-
-        [HttpGet("EMA")]
-        public IEnumerable<EmaResult> GetEMA(int lookbackPeriods)
+        catch (ArgumentOutOfRangeException rex)
         {
-            return quotes.GetEma(lookbackPeriods)
-                .Where(x => x.Date >= dateStart);
+            return BadRequest(rex.Message);
         }
+    }
 
-        [HttpGet("PSAR")]
-        public IEnumerable<ParabolicSarResult> GetParabolicSar(
-             decimal accelerationStep, decimal maxAccelerationFactor)
+    [HttpGet("PSAR")]
+    public IActionResult GetParabolicSar(
+         decimal accelerationStep,
+         decimal maxAccelerationFactor)
+    {
+        try
         {
-            return quotes.GetParabolicSar(accelerationStep, maxAccelerationFactor)
-                .Where(x => x.Date >= dateStart);
+            IEnumerable<ParabolicSarResult> results =
+                quotes.GetParabolicSar(accelerationStep, maxAccelerationFactor)
+                      .Where(x => x.Date >= dateStart);
+
+            return Ok(results);
         }
-
-        [HttpGet("RSI")]
-        public IEnumerable<RsiResult> GetRsi(int lookbackPeriods)
+        catch (ArgumentOutOfRangeException rex)
         {
-            return quotes.GetRsi(lookbackPeriods)
-                .Where(x => x.Date >= dateStart);
+            return BadRequest(rex.Message);
         }
+    }
 
-        [HttpGet("STOCH")]
-        public IEnumerable<StochResult> GetStoch(
-             int lookbackPeriods, int signalPeriods)
+    [HttpGet("RSI")]
+    public IActionResult GetRsi(int lookbackPeriods)
+    {
+        try
         {
-            return quotes.GetStoch(lookbackPeriods, signalPeriods)
-                .Where(x => x.Date >= dateStart);
+            IEnumerable<RsiResult> results =
+                quotes.GetRsi(lookbackPeriods)
+                      .Where(x => x.Date >= dateStart);
+
+            return Ok(results);
+        }
+        catch (ArgumentOutOfRangeException rex)
+        {
+            return BadRequest(rex.Message);
+        }
+    }
+
+    [HttpGet("STOCH")]
+    public IActionResult GetStoch(
+         int lookbackPeriods,
+         int signalPeriods)
+    {
+        try
+        {
+            IEnumerable<StochResult> results =
+                quotes.GetStoch(lookbackPeriods, signalPeriods)
+                      .Where(x => x.Date >= dateStart);
+
+            return Ok(results);
+        }
+        catch (ArgumentOutOfRangeException rex)
+        {
+            return BadRequest(rex.Message);
         }
     }
 }
