--- Server/WebApi/Services/QuotesService.cs
@@ -1,523 +1,0 @@
-﻿using Skender.Stock.Indicators;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-
-namespace WebApi.Services
-{
-    internal static class FetchQuotes
-    {
-        internal static IEnumerable<Quote> Get()
-        {
-            List<Quote> h = new()
-            {
-                new Quote { Date = DateTime.Parse("2018-12-31"), Open = (decimal)244.92, High = (decimal)245.54, Low = (decimal)242.87, Close = (decimal)245.28, Volume = 147031456 },
-                new Quote { Date = DateTime.Parse("2018-12-28"), Open = (decimal)244.94, High = (decimal)246.73, Low = (decimal)241.87, Close = (decimal)243.15, Volume = 155998912 },
-                new Quote { Date = DateTime.Parse("2018-12-27"), Open = (decimal)238.06, High = (decimal)243.68, Low = (decimal)234.52, Close = (decimal)243.46, Volume = 189794032 },
-                new Quote { Date = DateTime.Parse("2018-12-26"), Open = (decimal)231.59, High = (decimal)241.61, Low = (decimal)229.42, Close = (decimal)241.61, Volume = 222622048 },
-                new Quote { Date = DateTime.Parse("2018-12-24"), Open = (decimal)234.6, High = (decimal)236.36, Low = (decimal)229.92, Close = (decimal)229.99, Volume = 150100704 },
-                new Quote { Date = DateTime.Parse("2018-12-21"), Open = (decimal)242.16, High = (decimal)245.07, Low = (decimal)235.52, Close = (decimal)236.23, Volume = 260180208 },
-                new Quote { Date = DateTime.Parse("2018-12-20"), Open = (decimal)243.79, High = (decimal)245.51, Low = (decimal)238.71, Close = (decimal)241.17, Volume = 258325808 },
-                new Quote { Date = DateTime.Parse("2018-12-19"), Open = (decimal)248.97, High = (decimal)253.1, Low = (decimal)243.3, Close = (decimal)245.16, Volume = 220342928 },
-                new Quote { Date = DateTime.Parse("2018-12-18"), Open = (decimal)250.95, High = (decimal)251.69, Low = (decimal)247.13, Close = (decimal)248.89, Volume = 137862544 },
-                new Quote { Date = DateTime.Parse("2018-12-17"), Open = (decimal)253.1, High = (decimal)254.32, Low = (decimal)247.37, Close = (decimal)249.16, Volume = 169610592 },
-                new Quote { Date = DateTime.Parse("2018-12-14"), Open = (decimal)256.58, High = (decimal)257.62, Low = (decimal)253.54, Close = (decimal)254.15, Volume = 119871688 },
-                new Quote { Date = DateTime.Parse("2018-12-13"), Open = (decimal)260.05, High = (decimal)260.99, Low = (decimal)257.71, Close = (decimal)258.93, Volume = 99068200 },
-                new Quote { Date = DateTime.Parse("2018-12-12"), Open = (decimal)260.98, High = (decimal)262.47, Low = (decimal)258.93, Close = (decimal)259.01, Volume = 100414888 },
-                new Quote { Date = DateTime.Parse("2018-12-11"), Open = (decimal)261.16, High = (decimal)261.37, Low = (decimal)256.11, Close = (decimal)257.72, Volume = 124528112 },
-                new Quote { Date = DateTime.Parse("2018-12-10"), Open = (decimal)256.98, High = (decimal)258.72, Low = (decimal)252.34, Close = (decimal)257.66, Volume = 155214672 },
-                new Quote { Date = DateTime.Parse("2018-12-07"), Open = (decimal)262.92, High = (decimal)264.63, Low = (decimal)256.25, Close = (decimal)257.17, Volume = 165025936 },
-                new Quote { Date = DateTime.Parse("2018-12-06"), Open = (decimal)259.46, High = (decimal)263.41, Low = (decimal)256.07, Close = (decimal)263.29, Volume = 209266640 },
-                new Quote { Date = DateTime.Parse("2018-12-04"), Open = (decimal)271.61, High = (decimal)272.08, Low = (decimal)263.35, Close = (decimal)263.69, Volume = 182415248 },
-                new Quote { Date = DateTime.Parse("2018-12-03"), Open = (decimal)273.47, High = (decimal)273.59, Low = (decimal)270.77, Close = (decimal)272.52, Volume = 105581352 },
-                new Quote { Date = DateTime.Parse("2018-11-30"), Open = (decimal)267.16, High = (decimal)269.57, Low = (decimal)266.81, Close = (decimal)268.96, Volume = 100648032 },
-                new Quote { Date = DateTime.Parse("2018-11-29"), Open = (decimal)267.06, High = (decimal)268.86, Low = (decimal)265.82, Close = (decimal)267.33, Volume = 84395640 },
-                new Quote { Date = DateTime.Parse("2018-11-28"), Open = (decimal)263.05, High = (decimal)267.91, Low = (decimal)261.81, Close = (decimal)267.91, Volume = 130805744 },
-                new Quote { Date = DateTime.Parse("2018-11-27"), Open = (decimal)259.87, High = (decimal)261.88, Low = (decimal)259.21, Close = (decimal)261.88, Volume = 77381344 },
-                new Quote { Date = DateTime.Parse("2018-11-26"), Open = (decimal)259.33, High = (decimal)261.25, Low = (decimal)258.9, Close = 261, Volume = 81971728 },
-                new Quote { Date = DateTime.Parse("2018-11-23"), Open = (decimal)256.79, High = (decimal)258.39, Low = (decimal)256.68, Close = (decimal)256.86, Volume = 43873168 },
-                new Quote { Date = DateTime.Parse("2018-11-21"), Open = (decimal)259.4, High = (decimal)260.66, Low = (decimal)258.58, Close = (decimal)258.58, Volume = 77444168 },
-                new Quote { Date = DateTime.Parse("2018-11-20"), Open = (decimal)258.92, High = (decimal)260.52, Low = (decimal)256.76, Close = (decimal)257.71, Volume = 139406240 },
-                new Quote { Date = DateTime.Parse("2018-11-19"), Open = (decimal)266.42, High = (decimal)266.74, Low = (decimal)261.56, Close = (decimal)262.57, Volume = 105626432 },
-                new Quote { Date = DateTime.Parse("2018-11-16"), Open = (decimal)265.19, High = (decimal)268.08, Low = (decimal)264.62, Close = (decimal)267.08, Volume = 129820216 },
-                new Quote { Date = DateTime.Parse("2018-11-15"), Open = (decimal)262.25, High = (decimal)266.9, Low = (decimal)260.53, Close = (decimal)266.39, Volume = 138463488 },
-                new Quote { Date = DateTime.Parse("2018-11-14"), Open = (decimal)267.5, High = (decimal)267.94, Low = (decimal)261.93, Close = (decimal)263.64, Volume = 128454960 },
-                new Quote { Date = DateTime.Parse("2018-11-13"), Open = (decimal)266.46, High = (decimal)268.64, Low = (decimal)264.66, Close = (decimal)265.45, Volume = 100619768 },
-                new Quote { Date = DateTime.Parse("2018-11-12"), Open = (decimal)270.46, High = (decimal)270.72, Low = (decimal)265.39, Close = (decimal)265.95, Volume = 102153984 },
-                new Quote { Date = DateTime.Parse("2018-11-09"), Open = (decimal)272.25, High = (decimal)272.46, Low = (decimal)269.47, Close = (decimal)271.02, Volume = 101271544 },
-                new Quote { Date = DateTime.Parse("2018-11-08"), Open = (decimal)273.31, High = (decimal)274.39, Low = (decimal)272.44, Close = (decimal)273.69, Volume = 67216992 },
-                new Quote { Date = DateTime.Parse("2018-11-07"), Open = (decimal)270.82, High = (decimal)274.27, Low = (decimal)270.35, Close = (decimal)274.19, Volume = 105309120 },
-                new Quote { Date = DateTime.Parse("2018-11-06"), Open = (decimal)266.68, High = (decimal)268.62, Low = (decimal)266.62, Close = (decimal)268.44, Volume = 61581152 },
-                new Quote { Date = DateTime.Parse("2018-11-05"), Open = (decimal)265.82, High = (decimal)267.36, Low = (decimal)264.76, Close = (decimal)266.75, Volume = 67255520 },
-                new Quote { Date = DateTime.Parse("2018-11-02"), Open = (decimal)268.08, High = (decimal)268.55, Low = (decimal)263.04, Close = (decimal)265.29, Volume = 125685896 },
-                new Quote { Date = DateTime.Parse("2018-11-01"), Open = (decimal)265.01, High = (decimal)267.08, Low = (decimal)263.81, Close = (decimal)266.87, Volume = 101971008 },
-                new Quote { Date = DateTime.Parse("2018-10-31"), Open = (decimal)264.08, High = (decimal)266.6, Low = (decimal)263.56, Close = (decimal)264.06, Volume = 131489024 },
-                new Quote { Date = DateTime.Parse("2018-10-30"), Open = (decimal)257.27, High = (decimal)261.61, Low = (decimal)256.73, Close = (decimal)261.27, Volume = 161025888 },
-                new Quote { Date = DateTime.Parse("2018-10-29"), Open = (decimal)262.27, High = (decimal)263.69, Low = (decimal)253.54, Close = (decimal)257.45, Volume = 164749392 },
-                new Quote { Date = DateTime.Parse("2018-10-26"), Open = (decimal)259.46, High = (decimal)264.42, Low = (decimal)255.92, Close = (decimal)258.89, Volume = 206590848 },
-                new Quote { Date = DateTime.Parse("2018-10-25"), Open = (decimal)260.89, High = (decimal)265.21, Low = (decimal)259.77, Close = (decimal)263.52, Volume = 141497248 },
-                new Quote { Date = DateTime.Parse("2018-10-24"), Open = (decimal)266.69, High = (decimal)267.11, Low = (decimal)258.27, Close = (decimal)258.88, Volume = 182231472 },
-                new Quote { Date = DateTime.Parse("2018-10-23"), Open = (decimal)264.37, High = (decimal)268.2, Low = (decimal)262.09, Close = (decimal)266.97, Volume = 149994752 },
-                new Quote { Date = DateTime.Parse("2018-10-22"), Open = (decimal)270.27, High = (decimal)270.63, Low = (decimal)267.75, Close = (decimal)268.33, Volume = 84466760 },
-                new Quote { Date = DateTime.Parse("2018-10-19"), Open = (decimal)270.4, High = (decimal)272.52, Low = (decimal)268.78, Close = (decimal)269.54, Volume = 143383136 },
-                new Quote { Date = DateTime.Parse("2018-10-18"), Open = (decimal)272.62, High = (decimal)273.27, Low = (decimal)268.29, Close = (decimal)269.69, Volume = 137906032 },
-                new Quote { Date = DateTime.Parse("2018-10-17"), Open = (decimal)273.63, High = (decimal)274.32, Low = (decimal)270.82, Close = (decimal)273.64, Volume = 113378952 },
-                new Quote { Date = DateTime.Parse("2018-10-16"), Open = (decimal)269.88, High = 274, Low = (decimal)269.37, Close = (decimal)273.59, Volume = 121198672 },
-                new Quote { Date = DateTime.Parse("2018-10-15"), Open = (decimal)268.86, High = (decimal)270.31, Low = (decimal)267.64, Close = (decimal)267.74, Volume = 104808584 },
-                new Quote { Date = DateTime.Parse("2018-10-12"), Open = (decimal)270.05, High = (decimal)270.36, Low = (decimal)265.76, Close = (decimal)269.25, Volume = 187745152 },
-                new Quote { Date = DateTime.Parse("2018-10-11"), Open = (decimal)270.35, High = (decimal)272.13, Low = (decimal)263.8, Close = (decimal)265.56, Volume = 281680000 },
-                new Quote { Date = DateTime.Parse("2018-10-10"), Open = (decimal)279.87, High = (decimal)279.94, Low = (decimal)271.13, Close = (decimal)271.54, Volume = 220074688 },
-                new Quote { Date = DateTime.Parse("2018-10-09"), Open = (decimal)280.41, High = (decimal)281.85, Low = (decimal)279.81, Close = (decimal)280.42, Volume = 76188928 },
-                new Quote { Date = DateTime.Parse("2018-10-08"), Open = (decimal)280.08, High = (decimal)281.22, Low = (decimal)278.57, Close = (decimal)280.83, Volume = 89925664 },
-                new Quote { Date = DateTime.Parse("2018-10-05"), Open = (decimal)282.66, High = (decimal)283.22, Low = (decimal)279.27, Close = (decimal)280.83, Volume = 108588344 },
-                new Quote { Date = DateTime.Parse("2018-10-04"), Open = (decimal)284.11, High = (decimal)284.17, Low = (decimal)280.68, Close = (decimal)282.41, Volume = 114321768 },
-                new Quote { Date = DateTime.Parse("2018-10-03"), Open = (decimal)285.63, High = (decimal)286.09, Low = (decimal)284.25, Close = (decimal)284.64, Volume = 66304540 },
-                new Quote { Date = DateTime.Parse("2018-10-02"), Open = (decimal)284.48, High = (decimal)285.26, Low = (decimal)284.07, Close = (decimal)284.48, Volume = 48434264 },
-                new Quote { Date = DateTime.Parse("2018-10-01"), Open = (decimal)285.02, High = (decimal)285.82, Low = (decimal)283.91, Close = (decimal)284.65, Volume = 63623792 },
-                new Quote { Date = DateTime.Parse("2018-09-28"), Open = (decimal)282.95, High = (decimal)284.21, Low = (decimal)282.91, Close = (decimal)283.66, Volume = 71835632 },
-                new Quote { Date = DateTime.Parse("2018-09-27"), Open = (decimal)283.36, High = (decimal)284.82, Low = (decimal)283.06, Close = (decimal)283.63, Volume = 60723896 },
-                new Quote { Date = DateTime.Parse("2018-09-26"), Open = (decimal)283.85, High = (decimal)285.14, Low = (decimal)282.38, Close = (decimal)282.84, Volume = 81724024 },
-                new Quote { Date = DateTime.Parse("2018-09-25"), Open = (decimal)284.45, High = (decimal)284.57, Low = (decimal)283.43, Close = (decimal)283.69, Volume = 45474200 },
-                new Quote { Date = DateTime.Parse("2018-09-24"), Open = (decimal)284.27, High = (decimal)284.42, Low = (decimal)283.32, Close = (decimal)283.95, Volume = 54738760 },
-                new Quote { Date = DateTime.Parse("2018-09-21"), Open = (decimal)285.97, High = (decimal)286.1, Low = (decimal)284.72, Close = (decimal)284.9, Volume = 108104552 },
-                new Quote { Date = DateTime.Parse("2018-09-20"), Open = (decimal)284.25, High = (decimal)285.51, Low = (decimal)282.88, Close = (decimal)285.16, Volume = 103323632 },
-                new Quote { Date = DateTime.Parse("2018-09-19"), Open = (decimal)282.63, High = (decimal)283.33, Low = (decimal)282.48, Close = (decimal)282.87, Volume = 50529584 },
-                new Quote { Date = DateTime.Parse("2018-09-18"), Open = (decimal)281.28, High = (decimal)283.22, Low = (decimal)281.25, Close = (decimal)282.57, Volume = 63758804 },
-                new Quote { Date = DateTime.Parse("2018-09-17"), Open = (decimal)282.48, High = (decimal)282.52, Low = (decimal)280.74, Close = (decimal)281.04, Volume = 70258840 },
-                new Quote { Date = DateTime.Parse("2018-09-14"), Open = (decimal)282.71, High = (decimal)282.92, Low = (decimal)281.68, Close = (decimal)282.54, Volume = 56706020 },
-                new Quote { Date = DateTime.Parse("2018-09-13"), Open = (decimal)281.99, High = (decimal)282.69, Low = (decimal)281.68, Close = (decimal)282.49, Volume = 52540928 },
-                new Quote { Date = DateTime.Parse("2018-09-12"), Open = (decimal)280.77, High = (decimal)281.49, Low = (decimal)279.96, Close = (decimal)280.83, Volume = 61576576 },
-                new Quote { Date = DateTime.Parse("2018-09-11"), Open = (decimal)279.13, High = (decimal)281.25, Low = (decimal)278.75, Close = (decimal)280.76, Volume = 52022324 },
-                new Quote { Date = DateTime.Parse("2018-09-10"), Open = (decimal)280.46, High = (decimal)280.75, Low = (decimal)279.62, Close = (decimal)279.84, Volume = 51693300 },
-                new Quote { Date = DateTime.Parse("2018-09-07"), Open = (decimal)278.75, High = (decimal)280.42, Low = (decimal)278.49, Close = (decimal)279.35, Volume = 75695528 },
-                new Quote { Date = DateTime.Parse("2018-09-06"), Open = (decimal)280.86, High = (decimal)281.19, Low = (decimal)278.77, Close = (decimal)279.9, Volume = 67855744 },
-                new Quote { Date = DateTime.Parse("2018-09-05"), Open = (decimal)281.11, High = (decimal)281.33, Low = (decimal)279.63, Close = (decimal)280.74, Volume = 74591480 },
-                new Quote { Date = DateTime.Parse("2018-09-04"), Open = (decimal)281.53, High = (decimal)281.89, Low = (decimal)280.4, Close = (decimal)281.5, Volume = 59294748 },
-                new Quote { Date = DateTime.Parse("2018-08-31"), Open = (decimal)281.53, High = (decimal)282.47, Low = (decimal)280.99, Close = (decimal)281.98, Volume = 68093544 },
-                new Quote { Date = DateTime.Parse("2018-08-30"), Open = (decimal)282.6, High = 283, Low = (decimal)281.32, Close = (decimal)281.98, Volume = 63037208 },
-                new Quote { Date = DateTime.Parse("2018-08-29"), Open = (decimal)281.84, High = (decimal)283.37, Low = (decimal)281.57, Close = (decimal)283.12, Volume = 63300776 },
-                new Quote { Date = DateTime.Parse("2018-08-28"), Open = (decimal)281.98, High = (decimal)282.09, Low = (decimal)281.1, Close = (decimal)281.61, Volume = 48329404 },
-                new Quote { Date = DateTime.Parse("2018-08-27"), Open = (decimal)280.58, High = (decimal)281.59, Low = (decimal)280.4, Close = (decimal)281.47, Volume = 58757348 },
-                new Quote { Date = DateTime.Parse("2018-08-24"), Open = (decimal)278.23, High = (decimal)279.42, Low = (decimal)278.17, Close = (decimal)279.27, Volume = 59184624 },
-                new Quote { Date = DateTime.Parse("2018-08-23"), Open = (decimal)277.77, High = (decimal)278.71, Low = (decimal)277.24, Close = (decimal)277.59, Volume = 50657548 },
-                new Quote { Date = DateTime.Parse("2018-08-22"), Open = (decimal)277.68, High = (decimal)278.54, Low = (decimal)277.39, Close = (decimal)277.96, Volume = 46321688 },
-                new Quote { Date = DateTime.Parse("2018-08-21"), Open = (decimal)278.04, High = (decimal)279.07, Low = (decimal)277.52, Close = (decimal)278.13, Volume = 69258080 },
-                new Quote { Date = DateTime.Parse("2018-08-20"), Open = (decimal)277.38, High = (decimal)277.77, Low = (decimal)276.89, Close = (decimal)277.48, Volume = 40982744 },
-                new Quote { Date = DateTime.Parse("2018-08-17"), Open = (decimal)275.69, High = (decimal)277.37, Low = (decimal)275.24, Close = (decimal)276.89, Volume = 67555760 },
-                new Quote { Date = DateTime.Parse("2018-08-16"), Open = (decimal)275.27, High = (decimal)276.87, Low = (decimal)275.23, Close = (decimal)275.91, Volume = 72033608 },
-                new Quote { Date = DateTime.Parse("2018-08-15"), Open = (decimal)274.28, High = (decimal)274.44, Low = (decimal)272.13, Close = (decimal)273.7, Volume = 105964064 },
-                new Quote { Date = DateTime.Parse("2018-08-14"), Open = (decimal)274.81, High = (decimal)276.02, Low = (decimal)274.38, Close = (decimal)275.76, Volume = 45136396 },
-                new Quote { Date = DateTime.Parse("2018-08-13"), Open = (decimal)275.34, High = (decimal)276.01, Low = (decimal)273.69, Close = (decimal)274.01, Volume = 67673568 },
-                new Quote { Date = DateTime.Parse("2018-08-10"), Open = (decimal)275.32, High = (decimal)275.91, Low = (decimal)274.26, Close = (decimal)275.04, Volume = 79351592 },
-                new Quote { Date = DateTime.Parse("2018-08-09"), Open = (decimal)277.34, High = (decimal)277.77, Low = (decimal)276.74, Close = (decimal)276.9, Volume = 36771464 },
-                new Quote { Date = DateTime.Parse("2018-08-08"), Open = (decimal)277.21, High = (decimal)277.71, Low = (decimal)276.77, Close = (decimal)277.27, Volume = 43357916 },
-                new Quote { Date = DateTime.Parse("2018-08-07"), Open = (decimal)277.21, High = (decimal)277.81, Low = (decimal)277.06, Close = (decimal)277.39, Volume = 44471960 },
-                new Quote { Date = DateTime.Parse("2018-08-06"), Open = (decimal)275.51, High = (decimal)276.82, Low = (decimal)275.08, Close = (decimal)276.48, Volume = 40564136 },
-                new Quote { Date = DateTime.Parse("2018-08-03"), Open = (decimal)274.43, High = (decimal)275.52, Low = (decimal)274.23, Close = (decimal)275.47, Volume = 55527740 },
-                new Quote { Date = DateTime.Parse("2018-08-02"), Open = (decimal)271.38, High = (decimal)274.48, Low = (decimal)271.15, Close = (decimal)274.29, Volume = 65298924 },
-                new Quote { Date = DateTime.Parse("2018-08-01"), Open = (decimal)273.49, High = (decimal)274.04, Low = (decimal)272.1, Close = (decimal)272.81, Volume = 55443260 },
-                new Quote { Date = DateTime.Parse("2018-07-31"), Open = (decimal)272.76, High = (decimal)273.93, Low = (decimal)272.34, Close = (decimal)273.26, Volume = 70594928 },
-                new Quote { Date = DateTime.Parse("2018-07-30"), Open = (decimal)273.44, High = (decimal)273.61, Low = (decimal)271.35, Close = (decimal)271.92, Volume = 65624404 },
-                new Quote { Date = DateTime.Parse("2018-07-27"), Open = (decimal)275.57, High = (decimal)275.68, Low = (decimal)272.34, Close = (decimal)273.35, Volume = 79050080 },
-                new Quote { Date = DateTime.Parse("2018-07-26"), Open = (decimal)275.08, High = (decimal)275.96, Low = (decimal)274.97, Close = (decimal)275.21, Volume = 59629476 },
-                new Quote { Date = DateTime.Parse("2018-07-25"), Open = (decimal)273.26, High = (decimal)276.22, Low = (decimal)273.21, Close = (decimal)275.87, Volume = 81211824 },
-                new Quote { Date = DateTime.Parse("2018-07-24"), Open = (decimal)273.71, High = (decimal)274.46, Low = (decimal)272.58, Close = (decimal)273.53, Volume = 70035320 },
-                new Quote { Date = DateTime.Parse("2018-07-23"), Open = (decimal)271.44, High = (decimal)272.39, Low = (decimal)271.06, Close = (decimal)272.16, Volume = 48436568 },
-                new Quote { Date = DateTime.Parse("2018-07-20"), Open = (decimal)271.75, High = (decimal)272.44, Low = (decimal)271.48, Close = (decimal)271.66, Volume = 84804656 },
-                new Quote { Date = DateTime.Parse("2018-07-19"), Open = (decimal)272.27, High = (decimal)272.69, Low = (decimal)271.45, Close = (decimal)271.97, Volume = 63225212 },
-                new Quote { Date = DateTime.Parse("2018-07-18"), Open = (decimal)272.51, High = (decimal)273.12, Low = (decimal)272.03, Close = 273, Volume = 45910016 },
-                new Quote { Date = DateTime.Parse("2018-07-17"), Open = (decimal)270.48, High = (decimal)272.85, Low = (decimal)270.43, Close = (decimal)272.43, Volume = 53860032 },
-                new Quote { Date = DateTime.Parse("2018-07-16"), Open = (decimal)271.62, High = (decimal)271.78, Low = (decimal)270.84, Close = (decimal)271.33, Volume = 49624096 },
-                new Quote { Date = DateTime.Parse("2018-07-13"), Open = (decimal)271.16, High = (decimal)271.9, Low = (decimal)270.67, Close = (decimal)271.57, Volume = 49659024 },
-                new Quote { Date = DateTime.Parse("2018-07-12"), Open = (decimal)270.3, High = (decimal)271.42, Low = (decimal)269.64, Close = (decimal)271.36, Volume = 61899772 },
-                new Quote { Date = DateTime.Parse("2018-07-11"), Open = (decimal)269.2, High = (decimal)270.07, Low = (decimal)268.59, Close = (decimal)268.92, Volume = 79329656 },
-                new Quote { Date = DateTime.Parse("2018-07-10"), Open = (decimal)270.43, High = (decimal)271.01, Low = (decimal)270.11, Close = (decimal)270.9, Volume = 53501064 },
-                new Quote { Date = DateTime.Parse("2018-07-09"), Open = (decimal)268.62, High = (decimal)269.99, Low = (decimal)268.57, Close = (decimal)269.93, Volume = 52042820 },
-                new Quote { Date = DateTime.Parse("2018-07-06"), Open = (decimal)265.31, High = (decimal)267.93, Low = (decimal)264.89, Close = (decimal)267.52, Volume = 68456816 },
-                new Quote { Date = DateTime.Parse("2018-07-05"), Open = (decimal)264.36, High = (decimal)265.35, Low = (decimal)263.19, Close = (decimal)265.28, Volume = 58606568 },
-                new Quote { Date = DateTime.Parse("2018-07-03"), Open = (decimal)265.05, High = (decimal)265.15, Low = (decimal)262.67, Close = (decimal)263.13, Volume = 43432576 },
-                new Quote { Date = DateTime.Parse("2018-07-02"), Open = (decimal)261.78, High = (decimal)264.24, Low = (decimal)261.52, Close = (decimal)264.06, Volume = 65431128 },
-                new Quote { Date = DateTime.Parse("2018-06-29"), Open = (decimal)264.32, High = (decimal)265.81, Low = (decimal)263.37, Close = (decimal)263.5, Volume = 100473760 },
-                new Quote { Date = DateTime.Parse("2018-06-28"), Open = (decimal)261.57, High = (decimal)263.96, Low = (decimal)260.79, Close = (decimal)263.12, Volume = 78913504 },
-                new Quote { Date = DateTime.Parse("2018-06-27"), Open = (decimal)264.45, High = (decimal)266.01, Low = (decimal)261.46, Close = (decimal)261.63, Volume = 108213904 },
-                new Quote { Date = DateTime.Parse("2018-06-26"), Open = (decimal)263.85, High = (decimal)264.74, Low = (decimal)263.02, Close = (decimal)263.81, Volume = 70710976 },
-                new Quote { Date = DateTime.Parse("2018-06-25"), Open = (decimal)265.6, High = (decimal)265.77, Low = (decimal)261.38, Close = (decimal)263.23, Volume = 141924096 },
-                new Quote { Date = DateTime.Parse("2018-06-22"), Open = (decimal)267.76, High = (decimal)267.88, Low = (decimal)266.62, Close = (decimal)266.86, Volume = 58283384 },
-                new Quote { Date = DateTime.Parse("2018-06-21"), Open = (decimal)268.05, High = (decimal)268.07, Low = (decimal)265.83, Close = (decimal)266.38, Volume = 73159376 },
-                new Quote { Date = DateTime.Parse("2018-06-20"), Open = (decimal)268.35, High = (decimal)268.78, Low = (decimal)267.69, Close = (decimal)268.06, Volume = 55373416 },
-                new Quote { Date = DateTime.Parse("2018-06-19"), Open = (decimal)266.14, High = (decimal)267.84, Low = (decimal)265.69, Close = (decimal)267.6, Volume = 100410976 },
-                new Quote { Date = DateTime.Parse("2018-06-18"), Open = (decimal)267.59, High = (decimal)268.77, Low = (decimal)267.07, Close = (decimal)268.63, Volume = 54479888 },
-                new Quote { Date = DateTime.Parse("2018-06-15"), Open = (decimal)268.67, High = (decimal)269.55, Low = (decimal)267.45, Close = (decimal)269.18, Volume = 123585600 },
-                new Quote { Date = DateTime.Parse("2018-06-14"), Open = (decimal)269.8, High = (decimal)270.11, Low = (decimal)268.88, Close = (decimal)269.53, Volume = 79730104 },
-                new Quote { Date = DateTime.Parse("2018-06-13"), Open = (decimal)269.97, High = (decimal)270.25, Low = (decimal)268.63, Close = (decimal)268.85, Volume = 81770464 },
-                new Quote { Date = DateTime.Parse("2018-06-12"), Open = (decimal)269.82, High = (decimal)270.11, Low = 269, Close = (decimal)269.71, Volume = 74798688 },
-                new Quote { Date = DateTime.Parse("2018-06-11"), Open = (decimal)269.25, High = (decimal)270.15, Low = (decimal)269.12, Close = (decimal)269.36, Volume = 60903392 },
-                new Quote { Date = DateTime.Parse("2018-06-08"), Open = (decimal)267.71, High = (decimal)269.06, Low = (decimal)267.53, Close = 269, Volume = 74602920 },
-                new Quote { Date = DateTime.Parse("2018-06-07"), Open = (decimal)268.77, High = (decimal)269.09, Low = (decimal)267.22, Close = (decimal)268.21, Volume = 75460928 },
-                new Quote { Date = DateTime.Parse("2018-06-06"), Open = (decimal)266.68, High = (decimal)268.36, Low = (decimal)266.01, Close = (decimal)268.24, Volume = 64874192 },
-                new Quote { Date = DateTime.Parse("2018-06-05"), Open = (decimal)265.97, High = (decimal)266.43, Low = (decimal)265.13, Close = (decimal)266.02, Volume = 52881036 },
-                new Quote { Date = DateTime.Parse("2018-06-04"), Open = (decimal)265.47, High = (decimal)266.1, Low = (decimal)265.2, Close = (decimal)265.82, Volume = 46934832 },
-                new Quote { Date = DateTime.Parse("2018-06-01"), Open = (decimal)263.42, High = (decimal)264.9, Low = (decimal)263.34, Close = (decimal)264.57, Volume = 73691520 },
-                new Quote { Date = DateTime.Parse("2018-05-31"), Open = (decimal)263.16, High = (decimal)263.49, Low = (decimal)261.33, Close = (decimal)261.99, Volume = 96713160 },
-                new Quote { Date = DateTime.Parse("2018-05-30"), Open = (decimal)261.57, High = (decimal)264.09, Low = (decimal)261.49, Close = (decimal)263.61, Volume = 72057608 },
-                new Quote { Date = DateTime.Parse("2018-05-29"), Open = (decimal)261.39, High = (decimal)262.22, Low = (decimal)258.92, Close = (decimal)260.14, Volume = 119866288 },
-                new Quote { Date = DateTime.Parse("2018-05-25"), Open = (decimal)263.16, High = (decimal)263.85, Low = (decimal)262.61, Close = (decimal)263.16, Volume = 58299660 },
-                new Quote { Date = DateTime.Parse("2018-05-24"), Open = (decimal)263.9, High = (decimal)264.2, Low = (decimal)261.84, Close = (decimal)263.79, Volume = 78640328 },
-                new Quote { Date = DateTime.Parse("2018-05-23"), Open = (decimal)262.22, High = (decimal)264.36, Low = (decimal)262.04, Close = (decimal)264.33, Volume = 66903156 },
-                new Quote { Date = DateTime.Parse("2018-05-22"), Open = (decimal)264.91, High = (decimal)265.2, Low = (decimal)263.25, Close = (decimal)263.61, Volume = 54774884 },
-                new Quote { Date = DateTime.Parse("2018-05-21"), Open = 264, High = (decimal)264.93, Low = (decimal)262.39, Close = (decimal)264.34, Volume = 60007156 },
-                new Quote { Date = DateTime.Parse("2018-05-18"), Open = (decimal)262.65, High = (decimal)263.05, Low = (decimal)261.98, Close = (decimal)262.37, Volume = 66565792 },
-                new Quote { Date = DateTime.Parse("2018-05-17"), Open = (decimal)262.96, High = (decimal)264.21, Low = (decimal)262.18, Close = (decimal)263.03, Volume = 58466824 },
-                new Quote { Date = DateTime.Parse("2018-05-16"), Open = (decimal)262.19, High = (decimal)263.75, Low = (decimal)262.16, Close = (decimal)263.25, Volume = 55784492 },
-                new Quote { Date = DateTime.Parse("2018-05-15"), Open = (decimal)262.62, High = (decimal)262.64, Low = (decimal)261.11, Close = (decimal)262.15, Volume = 90007968 },
-                new Quote { Date = DateTime.Parse("2018-05-14"), Open = (decimal)264.31, High = (decimal)265.03, Low = (decimal)263.37, Close = (decimal)263.97, Volume = 56661420 },
-                new Quote { Date = DateTime.Parse("2018-05-11"), Open = (decimal)263.17, High = (decimal)264.13, Low = (decimal)262.61, Close = (decimal)263.84, Volume = 61915812 },
-                new Quote { Date = DateTime.Parse("2018-05-10"), Open = (decimal)261.41, High = (decimal)263.4, Low = (decimal)261.3, Close = (decimal)263.04, Volume = 74524544 },
-                new Quote { Date = DateTime.Parse("2018-05-09"), Open = (decimal)258.84, High = (decimal)260.95, Low = (decimal)258.27, Close = (decimal)260.6, Volume = 61703432 },
-                new Quote { Date = DateTime.Parse("2018-05-08"), Open = (decimal)257.7, High = (decimal)258.5, Low = (decimal)256.4, Close = (decimal)258.11, Volume = 69804000 },
-                new Quote { Date = DateTime.Parse("2018-05-07"), Open = (decimal)258.08, High = (decimal)259.17, Low = (decimal)257.32, Close = (decimal)258.11, Volume = 57193284 },
-                new Quote { Date = DateTime.Parse("2018-05-04"), Open = (decimal)252.89, High = (decimal)257.98, Low = (decimal)252.53, Close = (decimal)257.24, Volume = 94336840 },
-                new Quote { Date = DateTime.Parse("2018-05-03"), Open = (decimal)253.6, High = (decimal)254.66, Low = (decimal)250.5, Close = (decimal)253.95, Volume = 140965808 },
-                new Quote { Date = DateTime.Parse("2018-05-02"), Open = (decimal)256.02, High = (decimal)256.91, Low = (decimal)254.08, Close = (decimal)254.51, Volume = 89317992 },
-                new Quote { Date = DateTime.Parse("2018-05-01"), Open = (decimal)255.16, High = (decimal)256.35, Low = (decimal)253.46, Close = (decimal)256.23, Volume = 76737024 },
-                new Quote { Date = DateTime.Parse("2018-04-30"), Open = (decimal)258.44, High = (decimal)259.04, Low = (decimal)255.7, Close = (decimal)255.78, Volume = 84988424 },
-                new Quote { Date = DateTime.Parse("2018-04-27"), Open = (decimal)258.18, High = (decimal)258.51, Low = (decimal)256.73, Close = (decimal)257.76, Volume = 59001736 },
-                new Quote { Date = DateTime.Parse("2018-04-26"), Open = (decimal)256.05, High = (decimal)258.42, Low = (decimal)255.56, Close = (decimal)257.52, Volume = 70044640 },
-                new Quote { Date = DateTime.Parse("2018-04-25"), Open = (decimal)254.23, High = (decimal)255.41, Low = (decimal)252.24, Close = (decimal)254.93, Volume = 107386584 },
-                new Quote { Date = DateTime.Parse("2018-04-24"), Open = (decimal)258.89, High = (decimal)259.13, Low = (decimal)252.65, Close = (decimal)254.3, Volume = 116739904 },
-                new Quote { Date = DateTime.Parse("2018-04-23"), Open = (decimal)258.44, High = (decimal)259.04, Low = (decimal)256.59, Close = (decimal)257.77, Volume = 67796416 },
-                new Quote { Date = DateTime.Parse("2018-04-20"), Open = (decimal)259.93, High = (decimal)260.18, Low = (decimal)256.84, Close = (decimal)257.81, Volume = 103366016 },
-                new Quote { Date = DateTime.Parse("2018-04-19"), Open = (decimal)260.75, High = (decimal)260.97, Low = (decimal)258.88, Close = (decimal)260.01, Volume = 80307456 },
-                new Quote { Date = DateTime.Parse("2018-04-18"), Open = (decimal)261.75, High = (decimal)262.34, Low = (decimal)260.96, Close = (decimal)261.46, Volume = 59260488 },
-                new Quote { Date = DateTime.Parse("2018-04-17"), Open = (decimal)260.44, High = (decimal)261.93, Low = (decimal)259.88, Close = (decimal)261.27, Volume = 66890592 },
-                new Quote { Date = DateTime.Parse("2018-04-16"), Open = (decimal)258.18, High = (decimal)259.34, Low = (decimal)257.29, Close = (decimal)258.5, Volume = 65570252 },
-                new Quote { Date = DateTime.Parse("2018-04-13"), Open = (decimal)258.58, High = (decimal)258.71, Low = (decimal)255.29, Close = (decimal)256.4, Volume = 87984192 },
-                new Quote { Date = DateTime.Parse("2018-04-12"), Open = (decimal)256.5, High = (decimal)258.18, Low = (decimal)256.31, Close = (decimal)257.15, Volume = 71242736 },
-                new Quote { Date = DateTime.Parse("2018-04-11"), Open = (decimal)254.77, High = (decimal)256.87, Low = (decimal)254.69, Close = (decimal)255.05, Volume = 94252208 },
-                new Quote { Date = DateTime.Parse("2018-04-10"), Open = (decimal)255.54, High = (decimal)257.26, Low = (decimal)254.3, Close = (decimal)256.4, Volume = 109178536 },
-                new Quote { Date = DateTime.Parse("2018-04-09"), Open = (decimal)252.74, High = (decimal)256.1, Low = (decimal)251.35, Close = (decimal)252.38, Volume = 109043264 },
-                new Quote { Date = DateTime.Parse("2018-04-06"), Open = (decimal)254.72, High = (decimal)256.36, Low = (decimal)249.48, Close = (decimal)251.14, Volume = 185650928 },
-                new Quote { Date = DateTime.Parse("2018-04-05"), Open = (decimal)256.78, High = (decimal)257.84, Low = (decimal)255.59, Close = (decimal)256.87, Volume = 85474776 },
-                new Quote { Date = DateTime.Parse("2018-04-04"), Open = (decimal)248.27, High = (decimal)255.63, Low = (decimal)248.13, Close = (decimal)254.86, Volume = 127939576 },
-                new Quote { Date = DateTime.Parse("2018-04-03"), Open = (decimal)250.32, High = (decimal)252.68, Low = (decimal)248.36, Close = (decimal)252.16, Volume = 124052768 },
-                new Quote { Date = DateTime.Parse("2018-04-02"), Open = (decimal)253.88, High = (decimal)254.44, Low = (decimal)246.26, Close = (decimal)248.97, Volume = 192647056 },
-                new Quote { Date = DateTime.Parse("2018-03-29"), Open = (decimal)252.5, High = (decimal)256.5, Low = (decimal)251.26, Close = (decimal)254.46, Volume = 128487112 },
-                new Quote { Date = DateTime.Parse("2018-03-28"), Open = (decimal)252.14, High = (decimal)253.97, Low = (decimal)250.04, Close = (decimal)251.25, Volume = 151452896 },
-                new Quote { Date = DateTime.Parse("2018-03-27"), Open = (decimal)257.38, High = (decimal)257.96, Low = (decimal)250.29, Close = 252, Volume = 134378272 },
-                new Quote { Date = DateTime.Parse("2018-03-26"), Open = (decimal)253.48, High = (decimal)256.67, Low = (decimal)250.84, Close = (decimal)256.36, Volume = 146803168 },
-                new Quote { Date = DateTime.Parse("2018-03-23"), Open = (decimal)255.45, High = (decimal)256.27, Low = (decimal)249.32, Close = (decimal)249.53, Volume = 189801520 },
-                new Quote { Date = DateTime.Parse("2018-03-22"), Open = (decimal)259.06, High = (decimal)259.99, Low = (decimal)254.66, Close = (decimal)254.96, Volume = 153866192 },
-                new Quote { Date = DateTime.Parse("2018-03-21"), Open = (decimal)261.96, High = (decimal)264.25, Low = (decimal)261.27, Close = (decimal)261.5, Volume = 81397104 },
-                new Quote { Date = DateTime.Parse("2018-03-20"), Open = (decimal)261.99, High = (decimal)262.7, Low = (decimal)261.26, Close = 262, Volume = 61797672 },
-                new Quote { Date = DateTime.Parse("2018-03-19"), Open = (decimal)264.32, High = (decimal)265.34, Low = (decimal)259.75, Close = (decimal)261.56, Volume = 112937344 },
-                new Quote { Date = DateTime.Parse("2018-03-16"), Open = (decimal)265.44, High = (decimal)266.3, Low = (decimal)265.09, Close = (decimal)265.15, Volume = 103769888 },
-                new Quote { Date = DateTime.Parse("2018-03-15"), Open = (decimal)265.71, High = (decimal)266.41, Low = (decimal)264.31, Close = (decimal)264.86, Volume = 86627344 },
-                new Quote { Date = DateTime.Parse("2018-03-14"), Open = (decimal)267.57, High = (decimal)267.77, Low = (decimal)264.54, Close = (decimal)265.15, Volume = 109949368 },
-                new Quote { Date = DateTime.Parse("2018-03-13"), Open = (decimal)269.52, High = (decimal)270.07, Low = (decimal)265.85, Close = (decimal)266.52, Volume = 95490048 },
-                new Quote { Date = DateTime.Parse("2018-03-12"), Open = (decimal)268.9, High = (decimal)269.59, Low = (decimal)267.83, Close = (decimal)268.25, Volume = 74678496 },
-                new Quote { Date = DateTime.Parse("2018-03-09"), Open = (decimal)265.53, High = (decimal)268.59, Low = (decimal)265.19, Close = (decimal)268.59, Volume = 117975584 },
-                new Quote { Date = DateTime.Parse("2018-03-08"), Open = (decimal)263.46, High = (decimal)264.13, Low = (decimal)262.37, Close = (decimal)263.99, Volume = 69462520 },
-                new Quote { Date = DateTime.Parse("2018-03-07"), Open = (decimal)260.45, High = (decimal)263.11, Low = (decimal)260.24, Close = (decimal)262.72, Volume = 90396808 },
-                new Quote { Date = DateTime.Parse("2018-03-06"), Open = (decimal)263.22, High = (decimal)263.31, Low = (decimal)261.18, Close = (decimal)262.82, Volume = 82245904 },
-                new Quote { Date = DateTime.Parse("2018-03-05"), Open = (decimal)257.86, High = (decimal)262.83, Low = (decimal)257.74, Close = (decimal)262.15, Volume = 101032888 },
-                new Quote { Date = DateTime.Parse("2018-03-02"), Open = 256, High = (decimal)259.77, Low = (decimal)255.05, Close = (decimal)259.16, Volume = 144408144 },
-                new Quote { Date = DateTime.Parse("2018-03-01"), Open = (decimal)261.4, High = (decimal)263.1, Low = (decimal)256.19, Close = (decimal)257.83, Volume = 183626128 },
-                new Quote { Date = DateTime.Parse("2018-02-28"), Open = (decimal)265.51, High = (decimal)266.01, Low = (decimal)261.29, Close = (decimal)261.63, Volume = 126575120 },
-                new Quote { Date = DateTime.Parse("2018-02-27"), Open = (decimal)267.86, High = (decimal)268.63, Low = (decimal)264.24, Close = (decimal)264.31, Volume = 102893264 },
-                new Quote { Date = DateTime.Parse("2018-02-26"), Open = (decimal)265.76, High = (decimal)267.76, Low = (decimal)265.11, Close = (decimal)267.65, Volume = 89802808 },
-                new Quote { Date = DateTime.Parse("2018-02-23"), Open = (decimal)261.77, High = (decimal)264.58, Low = (decimal)261.25, Close = (decimal)264.58, Volume = 96318072 },
-                new Quote { Date = DateTime.Parse("2018-02-22"), Open = (decimal)261.1, High = (decimal)262.98, Low = (decimal)259.7, Close = (decimal)260.43, Volume = 114742312 },
-                new Quote { Date = DateTime.Parse("2018-02-21"), Open = (decimal)261.87, High = (decimal)264.59, Low = (decimal)259.99, Close = (decimal)260.09, Volume = 102669592 },
-                new Quote { Date = DateTime.Parse("2018-02-20"), Open = 262, High = (decimal)263.58, Low = (decimal)260.53, Close = (decimal)261.39, Volume = 89676400 },
-                new Quote { Date = DateTime.Parse("2018-02-16"), Open = (decimal)262.28, High = (decimal)265.17, Low = (decimal)262.23, Close = (decimal)263.04, Volume = 166561968 },
-                new Quote { Date = DateTime.Parse("2018-02-15"), Open = (decimal)261.56, High = (decimal)262.97, Low = (decimal)258.86, Close = (decimal)262.96, Volume = 115457688 },
-                new Quote { Date = DateTime.Parse("2018-02-14"), Open = (decimal)254.56, High = (decimal)260.04, Low = (decimal)254.55, Close = (decimal)259.65, Volume = 125358160 },
-                new Quote { Date = DateTime.Parse("2018-02-13"), Open = (decimal)254.24, High = (decimal)256.79, Low = (decimal)253.6, Close = (decimal)256.19, Volume = 84333360 },
-                new Quote { Date = DateTime.Parse("2018-02-12"), Open = (decimal)254.1, High = (decimal)257.16, Low = (decimal)252.02, Close = (decimal)255.56, Volume = 149239040 },
-                new Quote { Date = DateTime.Parse("2018-02-09"), Open = (decimal)251.18, High = (decimal)253.89, Low = (decimal)243.59, Close = (decimal)251.86, Volume = 294421856 },
-                new Quote { Date = DateTime.Parse("2018-02-08"), Open = (decimal)258.13, High = (decimal)258.28, Low = (decimal)248.09, Close = (decimal)248.13, Volume = 255885040 },
-                new Quote { Date = DateTime.Parse("2018-02-07"), Open = (decimal)258.6, High = (decimal)262.32, Low = (decimal)257.71, Close = (decimal)257.8, Volume = 173784240 },
-                new Quote { Date = DateTime.Parse("2018-02-06"), Open = (decimal)250.35, High = (decimal)259.76, Low = (decimal)249.16, Close = (decimal)259.21, Volume = 368619296 },
-                new Quote { Date = DateTime.Parse("2018-02-05"), Open = (decimal)263.37, High = (decimal)265.68, Low = (decimal)253.6, Close = (decimal)254.2, Volume = 305963968 },
-                new Quote { Date = DateTime.Parse("2018-02-02"), Open = (decimal)269.75, High = (decimal)269.9, Low = (decimal)265.25, Close = (decimal)265.29, Volume = 179804944 },
-                new Quote { Date = DateTime.Parse("2018-02-01"), Open = (decimal)270.71, High = (decimal)272.62, Low = (decimal)270.33, Close = (decimal)271.2, Volume = 93552120 },
-                new Quote { Date = DateTime.Parse("2018-01-31"), Open = (decimal)272.3, High = (decimal)272.85, Low = (decimal)270.33, Close = (decimal)271.51, Volume = 123502168 },
-                new Quote { Date = DateTime.Parse("2018-01-30"), Open = (decimal)272.18, High = (decimal)274.24, Low = (decimal)270.85, Close = (decimal)271.37, Volume = 136842368 },
-                new Quote { Date = DateTime.Parse("2018-01-29"), Open = (decimal)275.39, High = (decimal)275.87, Low = (decimal)274.01, Close = (decimal)274.18, Volume = 93568600 },
-                new Quote { Date = DateTime.Parse("2018-01-26"), Open = (decimal)273.77, High = (decimal)276.06, Low = (decimal)273.49, Close = (decimal)276.01, Volume = 111868160 },
-                new Quote { Date = DateTime.Parse("2018-01-25"), Open = (decimal)273.68, High = (decimal)273.79, Low = (decimal)271.99, Close = (decimal)272.85, Volume = 87825816 },
-                new Quote { Date = DateTime.Parse("2018-01-24"), Open = (decimal)273.55, High = (decimal)274.2, Low = (decimal)271.45, Close = (decimal)272.74, Volume = 139977680 },
-                new Quote { Date = DateTime.Parse("2018-01-23"), Open = (decimal)272.31, High = (decimal)273.16, Low = (decimal)271.96, Close = (decimal)272.84, Volume = 100801672 },
-                new Quote { Date = DateTime.Parse("2018-01-22"), Open = (decimal)269.84, High = (decimal)272.27, Low = (decimal)269.78, Close = (decimal)272.27, Volume = 94818768 },
-                new Quote { Date = DateTime.Parse("2018-01-19"), Open = (decimal)269.48, High = (decimal)270.07, Low = (decimal)268.85, Close = (decimal)270.07, Volume = 146315344 },
-                new Quote { Date = DateTime.Parse("2018-01-18"), Open = (decimal)269.17, High = (decimal)269.64, Low = (decimal)268.31, Close = (decimal)268.85, Volume = 104584464 },
-                new Quote { Date = DateTime.Parse("2018-01-17"), Open = (decimal)267.78, High = (decimal)269.72, Low = (decimal)266.76, Close = (decimal)269.3, Volume = 117595008 },
-                new Quote { Date = DateTime.Parse("2018-01-16"), Open = (decimal)269.05, High = (decimal)269.76, Low = 266, Close = (decimal)266.76, Volume = 110634704 },
-                new Quote { Date = DateTime.Parse("2018-01-12"), Open = (decimal)266.23, High = (decimal)267.86, Low = (decimal)265.9, Close = (decimal)267.67, Volume = 94293048 },
-                new Quote { Date = DateTime.Parse("2018-01-11"), Open = (decimal)264.62, High = (decimal)265.94, Low = (decimal)264.44, Close = (decimal)265.94, Volume = 64749016 },
-                new Quote { Date = DateTime.Parse("2018-01-10"), Open = (decimal)263.59, High = (decimal)264.3, Low = (decimal)262.86, Close = (decimal)264.01, Volume = 72238032 },
-                new Quote { Date = DateTime.Parse("2018-01-09"), Open = (decimal)264.28, High = (decimal)265.1, Low = (decimal)263.97, Close = (decimal)264.42, Volume = 59445976 },
-                new Quote { Date = DateTime.Parse("2018-01-08"), Open = (decimal)263.23, High = (decimal)263.99, Low = (decimal)262.91, Close = (decimal)263.82, Volume = 59513708 },
-                new Quote { Date = DateTime.Parse("2018-01-05"), Open = (decimal)262.46, High = (decimal)263.47, Low = (decimal)261.92, Close = (decimal)263.34, Volume = 86721784 },
-                new Quote { Date = DateTime.Parse("2018-01-04"), Open = (decimal)261.2, High = (decimal)262.12, Low = (decimal)260.57, Close = (decimal)261.59, Volume = 83723648 },
-                new Quote { Date = DateTime.Parse("2018-01-03"), Open = (decimal)259.04, High = (decimal)260.66, Low = (decimal)259.04, Close = (decimal)260.5, Volume = 93518840 },
-                new Quote { Date = DateTime.Parse("2018-01-02"), Open = (decimal)257.96, High = (decimal)258.9, Low = (decimal)257.54, Close = (decimal)258.86, Volume = 89973440 },
-                new Quote { Date = DateTime.Parse("2017-12-29"), Open = (decimal)258.63, High = (decimal)258.65, Low = (decimal)256.81, Close = (decimal)257.02, Volume = 99683152 },
-                new Quote { Date = DateTime.Parse("2017-12-28"), Open = (decimal)258.01, High = (decimal)258.04, Low = (decimal)257.59, Close = (decimal)257.99, Volume = 46843448 },
-                new Quote { Date = DateTime.Parse("2017-12-27"), Open = (decimal)257.52, High = (decimal)257.86, Low = (decimal)257.16, Close = (decimal)257.46, Volume = 59962032 },
-                new Quote { Date = DateTime.Parse("2017-12-26"), Open = (decimal)257.2, High = (decimal)257.58, Low = (decimal)257.04, Close = (decimal)257.34, Volume = 46976656 },
-                new Quote { Date = DateTime.Parse("2017-12-22"), Open = (decimal)257.73, High = (decimal)257.77, Low = (decimal)257.06, Close = (decimal)257.65, Volume = 81734768 },
-                new Quote { Date = DateTime.Parse("2017-12-21"), Open = (decimal)257.87, High = (decimal)258.49, Low = (decimal)257.44, Close = (decimal)257.71, Volume = 69598728 },
-                new Quote { Date = DateTime.Parse("2017-12-20"), Open = (decimal)258.38, High = (decimal)258.44, Low = (decimal)256.86, Close = (decimal)257.18, Volume = 79690000 },
-                new Quote { Date = DateTime.Parse("2017-12-19"), Open = (decimal)258.58, High = (decimal)258.63, Low = (decimal)257.24, Close = (decimal)257.32, Volume = 85536976 },
-                new Quote { Date = DateTime.Parse("2017-12-18"), Open = (decimal)258.21, High = (decimal)258.7, Low = (decimal)258.1, Close = (decimal)258.31, Volume = 86856320 },
-                new Quote { Date = DateTime.Parse("2017-12-15"), Open = (decimal)255.66, High = (decimal)257.19, Low = (decimal)255.6, Close = (decimal)256.68, Volume = 150146832 },
-                new Quote { Date = DateTime.Parse("2017-12-14"), Open = (decimal)255.93, High = (decimal)256.06, Low = (decimal)254.51, Close = (decimal)254.56, Volume = 105055176 },
-                new Quote { Date = DateTime.Parse("2017-12-13"), Open = (decimal)255.9, High = (decimal)256.38, Low = (decimal)255.51, Close = (decimal)255.61, Volume = 107391488 },
-                new Quote { Date = DateTime.Parse("2017-12-12"), Open = (decimal)255.43, High = (decimal)256.15, Low = (decimal)255.22, Close = (decimal)255.64, Volume = 88909792 },
-                new Quote { Date = DateTime.Parse("2017-12-11"), Open = (decimal)254.49, High = (decimal)255.25, Low = (decimal)254.39, Close = (decimal)255.19, Volume = 86699232 },
-                new Quote { Date = DateTime.Parse("2017-12-08"), Open = (decimal)253.92, High = (decimal)254.43, Low = 253, Close = (decimal)254.42, Volume = 79901608 },
-                new Quote { Date = DateTime.Parse("2017-12-07"), Open = (decimal)252.1, High = (decimal)253.38, Low = (decimal)251.96, Close = (decimal)253.04, Volume = 80584848 },
-                new Quote { Date = DateTime.Parse("2017-12-06"), Open = (decimal)251.89, High = (decimal)252.71, Low = (decimal)251.74, Close = (decimal)252.24, Volume = 79207304 },
-                new Quote { Date = DateTime.Parse("2017-12-05"), Open = (decimal)253.38, High = (decimal)254.07, Low = (decimal)252.05, Close = (decimal)252.2, Volume = 81394656 },
-                new Quote { Date = DateTime.Parse("2017-12-04"), Open = (decimal)255.19, High = (decimal)255.65, Low = (decimal)253.05, Close = (decimal)253.11, Volume = 98140184 },
-                new Quote { Date = DateTime.Parse("2017-12-01"), Open = (decimal)253.7, High = (decimal)254.23, Low = (decimal)249.87, Close = (decimal)253.41, Volume = 171557392 },
-                new Quote { Date = DateTime.Parse("2017-11-30"), Open = (decimal)252.74, High = (decimal)254.94, Low = (decimal)252.66, Close = (decimal)253.94, Volume = 133469840 },
-                new Quote { Date = DateTime.Parse("2017-11-29"), Open = (decimal)252.03, High = (decimal)252.62, Low = (decimal)251.25, Close = (decimal)251.74, Volume = 80891176 },
-                new Quote { Date = DateTime.Parse("2017-11-28"), Open = (decimal)249.87, High = (decimal)251.92, Low = (decimal)249.77, Close = (decimal)251.89, Volume = 103286312 },
-                new Quote { Date = DateTime.Parse("2017-11-27"), Open = (decimal)249.53, High = (decimal)249.86, Low = (decimal)249.14, Close = (decimal)249.36, Volume = 54553804 },
-                new Quote { Date = DateTime.Parse("2017-11-24"), Open = (decimal)249.45, High = (decimal)249.6, Low = (decimal)249.29, Close = (decimal)249.48, Volume = 29070892 },
-                new Quote { Date = DateTime.Parse("2017-11-22"), Open = (decimal)249.14, High = (decimal)249.28, Low = (decimal)248.73, Close = (decimal)248.91, Volume = 46996584 },
-                new Quote { Date = DateTime.Parse("2017-11-21"), Open = (decimal)248.35, High = (decimal)249.33, Low = (decimal)247.47, Close = (decimal)249.13, Volume = 72192504 },
-                new Quote { Date = DateTime.Parse("2017-11-20"), Open = (decimal)247.36, High = (decimal)247.73, Low = (decimal)247.09, Close = (decimal)247.51, Volume = 50171324 },
-                new Quote { Date = DateTime.Parse("2017-11-17"), Open = (decimal)247.43, High = (decimal)247.79, Low = 247, Close = (decimal)247.09, Volume = 79059392 },
-                new Quote { Date = DateTime.Parse("2017-11-16"), Open = (decimal)246.76, High = (decimal)248.22, Low = (decimal)246.72, Close = (decimal)247.82, Volume = 70731712 },
-                new Quote { Date = DateTime.Parse("2017-11-15"), Open = (decimal)245.9, High = (decimal)246.48, Low = (decimal)244.95, Close = (decimal)245.73, Volume = 84334432 },
-                new Quote { Date = DateTime.Parse("2017-11-14"), Open = (decimal)246.66, High = (decimal)247.08, Low = (decimal)245.8, Close = (decimal)246.96, Volume = 63988136 },
-                new Quote { Date = DateTime.Parse("2017-11-13"), Open = (decimal)246.56, High = (decimal)247.79, Low = (decimal)246.52, Close = (decimal)247.54, Volume = 52418324 },
-                new Quote { Date = DateTime.Parse("2017-11-10"), Open = (decimal)246.96, High = (decimal)247.5, Low = (decimal)246.62, Close = (decimal)247.31, Volume = 62599644 },
-                new Quote { Date = DateTime.Parse("2017-11-09"), Open = (decimal)246.96, High = (decimal)247.6, Low = (decimal)245.65, Close = (decimal)247.39, Volume = 99230672 },
-                new Quote { Date = DateTime.Parse("2017-11-08"), Open = (decimal)247.67, High = (decimal)248.39, Low = (decimal)247.37, Close = (decimal)248.29, Volume = 52669760 },
-                new Quote { Date = DateTime.Parse("2017-11-07"), Open = (decimal)248.15, High = (decimal)248.52, Low = (decimal)247.31, Close = (decimal)247.86, Volume = 60008920 },
-                new Quote { Date = DateTime.Parse("2017-11-06"), Open = (decimal)247.51, High = (decimal)248.18, Low = (decimal)247.43, Close = (decimal)248.04, Volume = 51817160 },
-                new Quote { Date = DateTime.Parse("2017-11-03"), Open = 247, High = (decimal)247.7, Low = (decimal)246.55, Close = (decimal)247.65, Volume = 62187496 },
-                new Quote { Date = DateTime.Parse("2017-11-02"), Open = (decimal)246.66, High = (decimal)246.98, Low = (decimal)245.49, Close = (decimal)246.83, Volume = 58910404 },
-                new Quote { Date = DateTime.Parse("2017-11-01"), Open = (decimal)247.26, High = (decimal)247.63, Low = (decimal)246.33, Close = (decimal)246.73, Volume = 56565660 },
-                new Quote { Date = DateTime.Parse("2017-10-31"), Open = (decimal)246.44, High = (decimal)246.69, Low = (decimal)246.08, Close = (decimal)246.41, Volume = 62933720 },
-                new Quote { Date = DateTime.Parse("2017-10-30"), Open = (decimal)246.33, High = (decimal)246.84, Low = (decimal)245.7, Close = (decimal)246.02, Volume = 56652224 },
-                new Quote { Date = DateTime.Parse("2017-10-27"), Open = (decimal)245.76, High = (decimal)247.12, Low = (decimal)244.95, Close = (decimal)246.94, Volume = 89292576 },
-                new Quote { Date = DateTime.Parse("2017-10-26"), Open = (decimal)245.3, High = (decimal)245.59, Low = (decimal)244.81, Close = (decimal)244.94, Volume = 72840760 },
-                new Quote { Date = DateTime.Parse("2017-10-25"), Open = (decimal)245.48, High = (decimal)245.6, Low = (decimal)243.39, Close = (decimal)244.63, Volume = 108236672 },
-                new Quote { Date = DateTime.Parse("2017-10-24"), Open = (decimal)245.88, High = (decimal)246.1, Low = (decimal)245.45, Close = (decimal)245.84, Volume = 69853928 },
-                new Quote { Date = DateTime.Parse("2017-10-23"), Open = (decimal)246.72, High = (decimal)246.75, Low = (decimal)245.33, Close = (decimal)245.41, Volume = 66701640 },
-                new Quote { Date = DateTime.Parse("2017-10-20"), Open = (decimal)245.98, High = (decimal)246.4, Low = (decimal)245.09, Close = (decimal)246.37, Volume = 93063952 },
-                new Quote { Date = DateTime.Parse("2017-10-19"), Open = (decimal)244.18, High = (decimal)245.14, Low = (decimal)243.72, Close = (decimal)245.1, Volume = 64602432 },
-                new Quote { Date = DateTime.Parse("2017-10-18"), Open = (decimal)245.21, High = (decimal)245.26, Low = (decimal)244.83, Close = (decimal)245.04, Volume = 42670820 },
-                new Quote { Date = DateTime.Parse("2017-10-17"), Open = (decimal)244.57, High = (decimal)244.85, Low = (decimal)244.33, Close = (decimal)244.8, Volume = 32936836 },
-                new Quote { Date = DateTime.Parse("2017-10-16"), Open = (decimal)244.55, High = (decimal)244.84, Low = (decimal)244.18, Close = (decimal)244.63, Volume = 39887916 },
-                new Quote { Date = DateTime.Parse("2017-10-13"), Open = (decimal)244.48, High = (decimal)244.61, Low = 244, Close = (decimal)244.3, Volume = 57189416 },
-                new Quote { Date = DateTime.Parse("2017-10-12"), Open = (decimal)244.02, High = (decimal)244.41, Low = (decimal)243.74, Close = 244, Volume = 49116908 },
-                new Quote { Date = DateTime.Parse("2017-10-11"), Open = (decimal)243.88, High = (decimal)244.37, Low = (decimal)243.7, Close = (decimal)244.37, Volume = 49752644 },
-                new Quote { Date = DateTime.Parse("2017-10-10"), Open = (decimal)243.96, High = (decimal)244.4, Low = (decimal)243.37, Close = (decimal)243.98, Volume = 44934412 },
-                new Quote { Date = DateTime.Parse("2017-10-09"), Open = (decimal)243.99, High = (decimal)244.06, Low = (decimal)243.05, Close = (decimal)243.34, Volume = 37363944 },
-                new Quote { Date = DateTime.Parse("2017-10-06"), Open = (decimal)243.53, High = (decimal)244.06, Low = (decimal)243.25, Close = (decimal)243.74, Volume = 84161696 },
-                new Quote { Date = DateTime.Parse("2017-10-05"), Open = (decimal)242.95, High = (decimal)244.04, Low = (decimal)242.62, Close = (decimal)244.02, Volume = 66291980 },
-                new Quote { Date = DateTime.Parse("2017-10-04"), Open = (decimal)242.13, High = (decimal)242.85, Low = (decimal)242.01, Close = (decimal)242.58, Volume = 58392872 },
-                new Quote { Date = DateTime.Parse("2017-10-03"), Open = (decimal)241.91, High = (decimal)242.33, Low = (decimal)241.69, Close = (decimal)242.3, Volume = 69722704 },
-                new Quote { Date = DateTime.Parse("2017-10-02"), Open = (decimal)240.98, High = (decimal)241.78, Low = (decimal)240.8, Close = (decimal)241.78, Volume = 61596044 },
-                new Quote { Date = DateTime.Parse("2017-09-29"), Open = (decimal)239.88, High = (decimal)240.82, Low = (decimal)239.68, Close = (decimal)240.74, Volume = 89308704 },
-                new Quote { Date = DateTime.Parse("2017-09-28"), Open = (decimal)239.3, High = (decimal)239.98, Low = (decimal)239.2, Close = (decimal)239.89, Volume = 46730936 },
-                new Quote { Date = DateTime.Parse("2017-09-27"), Open = (decimal)239.44, High = (decimal)240.03, Low = (decimal)238.47, Close = (decimal)239.6, Volume = 84532616 },
-                new Quote { Date = DateTime.Parse("2017-09-26"), Open = 239, High = (decimal)239.27, Low = (decimal)238.41, Close = (decimal)238.68, Volume = 56439616 },
-                new Quote { Date = DateTime.Parse("2017-09-25"), Open = (decimal)238.74, High = (decimal)239.13, Low = (decimal)237.72, Close = (decimal)238.53, Volume = 59552032 },
-                new Quote { Date = DateTime.Parse("2017-09-22"), Open = (decimal)238.65, High = (decimal)239.2, Low = (decimal)238.62, Close = (decimal)239.02, Volume = 53446664 },
-                new Quote { Date = DateTime.Parse("2017-09-21"), Open = (decimal)239.44, High = (decimal)239.54, Low = (decimal)238.78, Close = (decimal)238.97, Volume = 50313136 },
-                new Quote { Date = DateTime.Parse("2017-09-20"), Open = (decimal)239.62, High = (decimal)239.74, Low = (decimal)238.52, Close = (decimal)239.61, Volume = 62171164 },
-                new Quote { Date = DateTime.Parse("2017-09-19"), Open = (decimal)239.56, High = (decimal)239.62, Low = (decimal)239.17, Close = (decimal)239.53, Volume = 49161788 },
-                new Quote { Date = DateTime.Parse("2017-09-18"), Open = (decimal)239.18, High = (decimal)239.67, Low = (decimal)238.87, Close = (decimal)239.29, Volume = 48250824 },
-                new Quote { Date = DateTime.Parse("2017-09-15"), Open = (decimal)238.3, High = (decimal)238.88, Low = (decimal)238.19, Close = (decimal)238.78, Volume = 99592680 },
-                new Quote { Date = DateTime.Parse("2017-09-14"), Open = (decimal)238.18, High = (decimal)238.68, Low = (decimal)237.99, Close = (decimal)238.46, Volume = 100101416 },
-                new Quote { Date = DateTime.Parse("2017-09-13"), Open = (decimal)238.11, High = (decimal)238.57, Low = (decimal)237.98, Close = (decimal)238.54, Volume = 62116640 },
-                new Quote { Date = DateTime.Parse("2017-09-12"), Open = (decimal)238.02, High = (decimal)238.46, Low = (decimal)237.82, Close = (decimal)238.42, Volume = 59670932 },
-                new Quote { Date = DateTime.Parse("2017-09-11"), Open = (decimal)236.51, High = (decimal)237.71, Low = (decimal)236.49, Close = (decimal)237.62, Volume = 74845424 },
-                new Quote { Date = DateTime.Parse("2017-09-08"), Open = (decimal)235.07, High = (decimal)235.62, Low = (decimal)234.85, Close = (decimal)235.11, Volume = 66946052 },
-                new Quote { Date = DateTime.Parse("2017-09-07"), Open = (decimal)235.75, High = (decimal)235.77, Low = (decimal)234.94, Close = (decimal)235.39, Volume = 60865172 },
-                new Quote { Date = DateTime.Parse("2017-09-06"), Open = (decimal)235.36, High = (decimal)235.78, Low = (decimal)234.78, Close = (decimal)235.42, Volume = 60741564 },
-                new Quote { Date = DateTime.Parse("2017-09-05"), Open = (decimal)235.76, High = (decimal)236.01, Low = (decimal)233.56, Close = (decimal)234.62, Volume = 95856440 },
-                new Quote { Date = DateTime.Parse("2017-09-01"), Open = (decimal)236.39, High = (decimal)236.78, Low = (decimal)236.15, Close = (decimal)236.31, Volume = 65031164 },
-                new Quote { Date = DateTime.Parse("2017-08-31"), Open = (decimal)235.25, High = (decimal)236.25, Low = (decimal)234.61, Close = (decimal)235.98, Volume = 108866560 },
-                new Quote { Date = DateTime.Parse("2017-08-30"), Open = (decimal)233.44, High = (decimal)234.87, Low = (decimal)233.24, Close = (decimal)234.57, Volume = 65056144 },
-                new Quote { Date = DateTime.Parse("2017-08-29"), Open = (decimal)231.76, High = (decimal)233.75, Low = (decimal)231.63, Close = (decimal)233.46, Volume = 53629680 },
-                new Quote { Date = DateTime.Parse("2017-08-28"), Open = (decimal)233.77, High = (decimal)233.8, Low = (decimal)232.74, Close = (decimal)233.2, Volume = 42544052 },
-                new Quote { Date = DateTime.Parse("2017-08-25"), Open = (decimal)233.51, High = (decimal)234.19, Low = (decimal)233.02, Close = (decimal)233.19, Volume = 67589040 },
-                new Quote { Date = DateTime.Parse("2017-08-24"), Open = (decimal)233.61, High = (decimal)233.78, Low = (decimal)232.41, Close = (decimal)232.64, Volume = 53216420 },
-                new Quote { Date = DateTime.Parse("2017-08-23"), Open = (decimal)232.97, High = (decimal)233.65, Low = (decimal)232.81, Close = (decimal)233.19, Volume = 52652352 },
-                new Quote { Date = DateTime.Parse("2017-08-22"), Open = (decimal)232.24, High = (decimal)234.2, Low = (decimal)232.22, Close = (decimal)234.03, Volume = 66219544 },
-                new Quote { Date = DateTime.Parse("2017-08-21"), Open = (decimal)231.36, High = (decimal)231.89, Low = (decimal)230.58, Close = (decimal)231.6, Volume = 68662792 },
-                new Quote { Date = DateTime.Parse("2017-08-18"), Open = (decimal)231.6, High = (decimal)232.83, Low = (decimal)230.94, Close = (decimal)231.42, Volume = 143417408 },
-                new Quote { Date = DateTime.Parse("2017-08-17"), Open = (decimal)234.79, High = (decimal)235.13, Low = (decimal)231.79, Close = (decimal)231.79, Volume = 134757072 },
-                new Quote { Date = DateTime.Parse("2017-08-16"), Open = (decimal)235.62, High = (decimal)236.06, Low = (decimal)234.99, Close = (decimal)235.46, Volume = 59481648 },
-                new Quote { Date = DateTime.Parse("2017-08-15"), Open = (decimal)235.49, High = (decimal)235.51, Low = (decimal)234.71, Close = (decimal)235.05, Volume = 57937020 },
-                new Quote { Date = DateTime.Parse("2017-08-14"), Open = (decimal)234.17, High = (decimal)235.31, Low = (decimal)234.13, Close = (decimal)235.07, Volume = 76866480 },
-                new Quote { Date = DateTime.Parse("2017-08-11"), Open = (decimal)232.67, High = (decimal)233.42, Low = (decimal)232.41, Close = (decimal)232.77, Volume = 78521472 },
-                new Quote { Date = DateTime.Parse("2017-08-10"), Open = (decimal)234.84, High = (decimal)234.98, Low = (decimal)232.37, Close = (decimal)232.42, Volume = 126355448 },
-                new Quote { Date = DateTime.Parse("2017-08-09"), Open = (decimal)235.01, High = (decimal)235.81, Low = (decimal)234.62, Close = (decimal)235.75, Volume = 65687312 },
-                new Quote { Date = DateTime.Parse("2017-08-08"), Open = 236, High = (decimal)237.33, Low = (decimal)235.35, Close = (decimal)235.76, Volume = 64729500 },
-                new Quote { Date = DateTime.Parse("2017-08-07"), Open = (decimal)235.98, High = (decimal)236.34, Low = (decimal)235.87, Close = (decimal)236.34, Volume = 33555464 },
-                new Quote { Date = DateTime.Parse("2017-08-04"), Open = (decimal)236.01, High = (decimal)236.27, Low = (decimal)235.49, Close = (decimal)235.9, Volume = 63127488 },
-                new Quote { Date = DateTime.Parse("2017-08-03"), Open = (decimal)235.81, High = (decimal)235.84, Low = (decimal)235.17, Close = (decimal)235.48, Volume = 42848608 },
-                new Quote { Date = DateTime.Parse("2017-08-02"), Open = (decimal)235.96, High = (decimal)236.09, Low = (decimal)234.91, Close = (decimal)235.93, Volume = 49513776 },
-                new Quote { Date = DateTime.Parse("2017-08-01"), Open = (decimal)235.95, High = (decimal)235.99, Low = (decimal)235.24, Close = (decimal)235.82, Volume = 57735292 },
-                new Quote { Date = DateTime.Parse("2017-07-31"), Open = (decimal)235.87, High = (decimal)235.97, Low = (decimal)235.07, Close = (decimal)235.29, Volume = 69049712 },
-                new Quote { Date = DateTime.Parse("2017-07-28"), Open = (decimal)235.18, High = (decimal)235.57, Low = (decimal)234.68, Close = (decimal)235.43, Volume = 52531244 },
-                new Quote { Date = DateTime.Parse("2017-07-27"), Open = (decimal)236.43, High = (decimal)236.47, Low = (decimal)234.26, Close = (decimal)235.7, Volume = 74217968 },
-                new Quote { Date = DateTime.Parse("2017-07-26"), Open = (decimal)236.23, High = (decimal)236.27, Low = (decimal)235.64, Close = (decimal)235.92, Volume = 49895744 },
-                new Quote { Date = DateTime.Parse("2017-07-25"), Open = (decimal)236.16, High = (decimal)236.28, Low = (decimal)235.67, Close = (decimal)235.91, Volume = 57593908 },
-                new Quote { Date = DateTime.Parse("2017-07-24"), Open = (decimal)235.31, High = (decimal)235.49, Low = (decimal)234.83, Close = (decimal)235.34, Volume = 48896096 },
-                new Quote { Date = DateTime.Parse("2017-07-21"), Open = (decimal)234.98, High = (decimal)235.43, Low = (decimal)234.73, Close = (decimal)235.4, Volume = 93037592 },
-                new Quote { Date = DateTime.Parse("2017-07-20"), Open = (decimal)235.78, High = (decimal)235.91, Low = (decimal)235.01, Close = (decimal)235.61, Volume = 49434036 },
-                new Quote { Date = DateTime.Parse("2017-07-19"), Open = (decimal)234.58, High = (decimal)235.51, Low = (decimal)234.57, Close = (decimal)235.5, Volume = 53523280 },
-                new Quote { Date = DateTime.Parse("2017-07-18"), Open = (decimal)233.66, High = (decimal)234.29, Low = (decimal)233.29, Close = (decimal)234.24, Volume = 44827112 },
-                new Quote { Date = DateTime.Parse("2017-07-17"), Open = (decimal)234.05, High = (decimal)234.47, Low = (decimal)233.92, Close = (decimal)234.11, Volume = 35167316 },
-                new Quote { Date = DateTime.Parse("2017-07-14"), Open = (decimal)233.06, High = (decimal)234.53, Low = (decimal)232.95, Close = (decimal)234.14, Volume = 63201796 },
-                new Quote { Date = DateTime.Parse("2017-07-13"), Open = (decimal)232.67, High = (decimal)233.18, Low = (decimal)232.42, Close = (decimal)233.05, Volume = 41396728 },
-                new Quote { Date = DateTime.Parse("2017-07-12"), Open = (decimal)231.99, High = (decimal)232.84, Low = (decimal)231.99, Close = (decimal)232.66, Volume = 62517696 },
-                new Quote { Date = DateTime.Parse("2017-07-11"), Open = (decimal)230.9, High = (decimal)231.27, Low = (decimal)229.65, Close = (decimal)230.93, Volume = 52810484 },
-                new Quote { Date = DateTime.Parse("2017-07-10"), Open = (decimal)230.7, High = (decimal)231.51, Low = (decimal)230.52, Close = (decimal)231.1, Volume = 38451396 },
-                new Quote { Date = DateTime.Parse("2017-07-07"), Open = (decimal)229.99, High = (decimal)231.01, Low = (decimal)229.38, Close = (decimal)230.85, Volume = 60799664 },
-                new Quote { Date = DateTime.Parse("2017-07-06"), Open = (decimal)230.64, High = (decimal)230.77, Low = (decimal)229.16, Close = (decimal)229.36, Volume = 69339864 },
-                new Quote { Date = DateTime.Parse("2017-07-05"), Open = (decimal)231.35, High = (decimal)231.71, Low = (decimal)230.46, Close = (decimal)231.48, Volume = 57082112 },
-                new Quote { Date = DateTime.Parse("2017-07-03"), Open = (decimal)231.59, High = (decimal)232.06, Low = (decimal)230.95, Close = (decimal)230.95, Volume = 41063396 },
-                new Quote { Date = DateTime.Parse("2017-06-30"), Open = (decimal)231.01, High = (decimal)231.42, Low = (decimal)230.34, Close = (decimal)230.56, Volume = 91055080 },
-                new Quote { Date = DateTime.Parse("2017-06-29"), Open = (decimal)232.33, High = (decimal)232.39, Low = (decimal)228.8, Close = (decimal)230.13, Volume = 112165824 },
-                new Quote { Date = DateTime.Parse("2017-06-28"), Open = (decimal)231.22, High = (decimal)232.38, Low = (decimal)230.97, Close = (decimal)232.17, Volume = 73458688 },
-                new Quote { Date = DateTime.Parse("2017-06-27"), Open = (decimal)231.74, High = (decimal)232.06, Low = (decimal)230.09, Close = (decimal)230.11, Volume = 86259016 },
-                new Quote { Date = DateTime.Parse("2017-06-26"), Open = (decimal)232.56, High = (decimal)233.02, Low = (decimal)231.74, Close = (decimal)231.98, Volume = 59465848 },
-                new Quote { Date = DateTime.Parse("2017-06-23"), Open = (decimal)231.61, High = (decimal)232.19, Low = (decimal)231.19, Close = (decimal)231.82, Volume = 70253848 },
-                new Quote { Date = DateTime.Parse("2017-06-22"), Open = (decimal)231.66, High = (decimal)232.21, Low = (decimal)231.36, Close = (decimal)231.55, Volume = 46301224 },
-                new Quote { Date = DateTime.Parse("2017-06-21"), Open = (decimal)232.1, High = (decimal)232.26, Low = (decimal)231.14, Close = (decimal)231.65, Volume = 58707680 },
-                new Quote { Date = DateTime.Parse("2017-06-20"), Open = (decimal)232.89, High = (decimal)232.9, Low = (decimal)231.69, Close = (decimal)231.71, Volume = 59681776 },
-                new Quote { Date = DateTime.Parse("2017-06-19"), Open = (decimal)232.26, High = (decimal)233.35, Low = (decimal)232.16, Close = (decimal)233.28, Volume = 68299992 },
-                new Quote { Date = DateTime.Parse("2017-06-16"), Open = (decimal)231.48, High = (decimal)231.54, Low = (decimal)230.4, Close = (decimal)231.36, Volume = 88676880 },
-                new Quote { Date = DateTime.Parse("2017-06-15"), Open = (decimal)230.27, High = (decimal)231.44, Low = (decimal)229.97, Close = (decimal)231.31, Volume = 70046440 },
-                new Quote { Date = DateTime.Parse("2017-06-14"), Open = (decimal)232.34, High = (decimal)232.35, Low = (decimal)230.85, Close = (decimal)231.75, Volume = 82837904 },
-                new Quote { Date = DateTime.Parse("2017-06-13"), Open = (decimal)231.51, High = (decimal)232.1, Low = (decimal)231.13, Close = (decimal)232.05, Volume = 63303744 },
-                new Quote { Date = DateTime.Parse("2017-06-12"), Open = (decimal)230.7, High = (decimal)230.97, Low = (decimal)229.99, Close = (decimal)230.92, Volume = 90748160 },
-                new Quote { Date = DateTime.Parse("2017-06-09"), Open = (decimal)231.61, High = (decimal)232.48, Low = (decimal)229.58, Close = (decimal)230.96, Volume = 139383184 },
-                new Quote { Date = DateTime.Parse("2017-06-08"), Open = (decimal)231.31, High = (decimal)231.84, Low = (decimal)230.74, Close = (decimal)231.32, Volume = 69504536 },
-                new Quote { Date = DateTime.Parse("2017-06-07"), Open = (decimal)231.14, High = (decimal)231.45, Low = (decimal)230.41, Close = (decimal)231.2, Volume = 57061952 },
-                new Quote { Date = DateTime.Parse("2017-06-06"), Open = (decimal)230.9, High = (decimal)231.51, Low = (decimal)230.69, Close = (decimal)230.77, Volume = 53089976 },
-                new Quote { Date = DateTime.Parse("2017-06-05"), Open = (decimal)231.5, High = (decimal)231.81, Low = (decimal)231.3, Close = (decimal)231.51, Volume = 47107480 },
-                new Quote { Date = DateTime.Parse("2017-06-02"), Open = (decimal)230.97, High = (decimal)231.86, Low = (decimal)230.65, Close = (decimal)231.69, Volume = 93444032 },
-                new Quote { Date = DateTime.Parse("2017-06-01"), Open = (decimal)229.6, High = (decimal)230.94, Low = (decimal)229.28, Close = (decimal)230.92, Volume = 72678144 },
-                new Quote { Date = DateTime.Parse("2017-05-31"), Open = (decimal)229.47, High = (decimal)229.51, Low = (decimal)228.34, Close = (decimal)229.09, Volume = 96742576 },
-                new Quote { Date = DateTime.Parse("2017-05-30"), Open = 229, High = (decimal)229.43, Low = (decimal)228.83, Close = (decimal)229.15, Volume = 37098796 },
-                new Quote { Date = DateTime.Parse("2017-05-26"), Open = (decimal)229.19, High = (decimal)229.53, Low = (decimal)229.1, Close = (decimal)229.35, Volume = 49142620 },
-                new Quote { Date = DateTime.Parse("2017-05-25"), Open = (decimal)228.87, High = (decimal)229.7, Low = (decimal)228.64, Close = (decimal)229.4, Volume = 67524256 },
-                new Quote { Date = DateTime.Parse("2017-05-24"), Open = (decimal)228.03, High = (decimal)228.42, Low = (decimal)227.66, Close = (decimal)228.31, Volume = 51831288 },
-                new Quote { Date = DateTime.Parse("2017-05-23"), Open = (decimal)227.68, High = (decimal)227.96, Low = (decimal)227.26, Close = (decimal)227.78, Volume = 50946640 },
-                new Quote { Date = DateTime.Parse("2017-05-22"), Open = (decimal)226.68, High = (decimal)227.45, Low = (decimal)226.61, Close = (decimal)227.27, Volume = 64298244 },
-                new Quote { Date = DateTime.Parse("2017-05-19"), Open = (decimal)225.2, High = (decimal)226.86, Low = (decimal)225.14, Close = (decimal)226.12, Volume = 121208928 },
-                new Quote { Date = DateTime.Parse("2017-05-18"), Open = (decimal)223.68, High = (decimal)225.59, Low = (decimal)223.39, Close = (decimal)224.66, Volume = 112816072 },
-                new Quote { Date = DateTime.Parse("2017-05-17"), Open = (decimal)225.93, High = (decimal)226.44, Low = (decimal)223.7, Close = (decimal)223.76, Volume = 181451968 },
-                new Quote { Date = DateTime.Parse("2017-05-16"), Open = (decimal)228.34, High = (decimal)228.36, Low = (decimal)227.38, Close = (decimal)227.8, Volume = 54003024 },
-                new Quote { Date = DateTime.Parse("2017-05-15"), Open = (decimal)227.23, High = (decimal)228.15, Low = (decimal)227.21, Close = (decimal)228.01, Volume = 65255528 },
-                new Quote { Date = DateTime.Parse("2017-05-12"), Open = (decimal)226.87, High = (decimal)227.19, Low = (decimal)226.47, Close = (decimal)226.76, Volume = 56817892 },
-                new Quote { Date = DateTime.Parse("2017-05-11"), Open = (decimal)227.11, High = (decimal)227.32, Low = (decimal)225.95, Close = (decimal)227.14, Volume = 65718612 },
-                new Quote { Date = DateTime.Parse("2017-05-10"), Open = (decimal)227.15, High = (decimal)227.61, Low = (decimal)226.92, Close = (decimal)227.61, Volume = 57219496 },
-                new Quote { Date = DateTime.Parse("2017-05-09"), Open = (decimal)227.69, High = (decimal)227.91, Low = (decimal)226.82, Close = (decimal)227.2, Volume = 54130976 },
-                new Quote { Date = DateTime.Parse("2017-05-08"), Open = (decimal)227.49, High = (decimal)227.65, Low = (decimal)226.94, Close = (decimal)227.41, Volume = 50993060 },
-                new Quote { Date = DateTime.Parse("2017-05-05"), Open = (decimal)226.96, High = (decimal)227.46, Low = (decimal)226.48, Close = (decimal)227.44, Volume = 65342296 },
-                new Quote { Date = DateTime.Parse("2017-05-04"), Open = (decimal)226.62, High = (decimal)226.71, Low = (decimal)225.62, Close = (decimal)226.55, Volume = 64774736 },
-                new Quote { Date = DateTime.Parse("2017-05-03"), Open = (decimal)226.11, High = (decimal)226.66, Low = (decimal)225.55, Close = (decimal)226.29, Volume = 77078864 },
-                new Quote { Date = DateTime.Parse("2017-05-02"), Open = (decimal)226.63, High = (decimal)226.76, Low = (decimal)226.12, Close = (decimal)226.56, Volume = 60467504 },
-                new Quote { Date = DateTime.Parse("2017-05-01"), Open = (decimal)226.48, High = (decimal)226.94, Low = (decimal)226.02, Close = (decimal)226.48, Volume = 70486576 },
-                new Quote { Date = DateTime.Parse("2017-04-28"), Open = (decimal)226.68, High = (decimal)226.71, Low = (decimal)225.76, Close = (decimal)225.91, Volume = 66956400 },
-                new Quote { Date = DateTime.Parse("2017-04-27"), Open = (decimal)226.56, High = (decimal)226.73, Low = (decimal)225.81, Close = (decimal)226.4, Volume = 60503960 },
-                new Quote { Date = DateTime.Parse("2017-04-26"), Open = (decimal)226.31, High = (decimal)227.28, Low = (decimal)226.16, Close = (decimal)226.21, Volume = 89266768 },
-                new Quote { Date = DateTime.Parse("2017-04-25"), Open = (decimal)225.75, High = (decimal)226.73, Low = (decimal)225.65, Close = (decimal)226.35, Volume = 80831256 },
-                new Quote { Date = DateTime.Parse("2017-04-24"), Open = (decimal)225.05, High = (decimal)225.27, Low = (decimal)222.57, Close = (decimal)225.04, Volume = 125633672 },
-                new Quote { Date = DateTime.Parse("2017-04-21"), Open = (decimal)223.22, High = (decimal)223.28, Low = (decimal)222.16, Close = (decimal)222.6, Volume = 116338368 },
-                new Quote { Date = DateTime.Parse("2017-04-20"), Open = (decimal)222.18, High = (decimal)223.79, Low = (decimal)221.83, Close = (decimal)223.31, Volume = 97560568 },
-                new Quote { Date = DateTime.Parse("2017-04-19"), Open = (decimal)222.53, High = (decimal)222.94, Low = (decimal)221.26, Close = (decimal)221.5, Volume = 72401856 },
-                new Quote { Date = DateTime.Parse("2017-04-18"), Open = (decimal)221.77, High = (decimal)222.5, Low = (decimal)221.16, Close = (decimal)221.91, Volume = 87710560 },
-                new Quote { Date = DateTime.Parse("2017-04-17"), Open = (decimal)221.19, High = (decimal)222.58, Low = (decimal)220.97, Close = (decimal)222.58, Volume = 72091488 },
-                new Quote { Date = DateTime.Parse("2017-04-13"), Open = (decimal)221.69, High = (decimal)222.5, Low = (decimal)220.62, Close = (decimal)220.62, Volume = 97885392 },
-                new Quote { Date = DateTime.Parse("2017-04-12"), Open = (decimal)222.74, High = (decimal)222.95, Low = (decimal)221.82, Close = (decimal)222.06, Volume = 86275816 },
-                new Quote { Date = DateTime.Parse("2017-04-11"), Open = (decimal)222.89, High = (decimal)223.15, Low = (decimal)221.41, Close = (decimal)223.04, Volume = 92789720 },
-                new Quote { Date = DateTime.Parse("2017-04-10"), Open = (decimal)223.33, High = (decimal)224.18, Low = (decimal)222.73, Close = (decimal)223.31, Volume = 71258848 },
-                new Quote { Date = DateTime.Parse("2017-04-07"), Open = (decimal)223.13, High = (decimal)223.93, Low = (decimal)222.64, Close = (decimal)223.17, Volume = 78422128 },
-                new Quote { Date = DateTime.Parse("2017-04-06"), Open = (decimal)222.93, High = (decimal)223.97, Low = (decimal)222.44, Close = (decimal)223.4, Volume = 72861232 },
-                new Quote { Date = DateTime.Parse("2017-04-05"), Open = (decimal)224.18, High = (decimal)225.25, Low = (decimal)222.55, Close = (decimal)222.78, Volume = 114663488 },
-                new Quote { Date = DateTime.Parse("2017-04-04"), Open = (decimal)222.98, High = (decimal)223.53, Low = (decimal)222.56, Close = (decimal)223.44, Volume = 59508952 },
-                new Quote { Date = DateTime.Parse("2017-04-03"), Open = (decimal)223.74, High = (decimal)223.96, Low = (decimal)221.95, Close = (decimal)223.3, Volume = 90156280 },
-                new Quote { Date = DateTime.Parse("2017-03-31"), Open = (decimal)223.84, High = (decimal)224.42, Low = (decimal)223.63, Close = (decimal)223.69, Volume = 77706304 },
-                new Quote { Date = DateTime.Parse("2017-03-30"), Open = (decimal)223.43, High = (decimal)224.43, Low = (decimal)223.24, Close = (decimal)224.21, Volume = 59795288 },
-                new Quote { Date = DateTime.Parse("2017-03-29"), Open = (decimal)222.97, High = (decimal)223.75, Low = (decimal)222.72, Close = (decimal)223.5, Volume = 65288636 },
-                new Quote { Date = DateTime.Parse("2017-03-28"), Open = (decimal)221.34, High = (decimal)223.75, Low = (decimal)221.22, Close = (decimal)223.29, Volume = 98521432 },
-                new Quote { Date = DateTime.Parse("2017-03-27"), Open = (decimal)220.07, High = (decimal)221.96, Low = (decimal)219.77, Close = (decimal)221.67, Volume = 92167056 },
-                new Quote { Date = DateTime.Parse("2017-03-24"), Open = (decimal)222.4, High = (decimal)223.02, Low = (decimal)221.05, Close = (decimal)221.9, Volume = 118567344 },
-                new Quote { Date = DateTime.Parse("2017-03-23"), Open = (decimal)222.04, High = (decimal)223.31, Low = (decimal)221.66, Close = (decimal)222.06, Volume = 105821032 },
-                new Quote { Date = DateTime.Parse("2017-03-22"), Open = (decimal)221.82, High = (decimal)222.61, Low = (decimal)221.13, Close = (decimal)222.3, Volume = 102826864 },
-                new Quote { Date = DateTime.Parse("2017-03-21"), Open = (decimal)225.33, High = (decimal)225.46, Low = (decimal)221.64, Close = (decimal)221.78, Volume = 138912016 },
-                new Quote { Date = DateTime.Parse("2017-03-20"), Open = (decimal)224.91, High = (decimal)225.22, Low = (decimal)224.24, Close = (decimal)224.66, Volume = 55368008 },
-                new Quote { Date = DateTime.Parse("2017-03-17"), Open = (decimal)225.59, High = (decimal)225.8, Low = (decimal)224.91, Close = (decimal)224.91, Volume = 93798120 },
-                new Quote { Date = DateTime.Parse("2017-03-16"), Open = (decimal)225.9, High = (decimal)225.99, Low = (decimal)224.95, Close = (decimal)225.31, Volume = 82924856 },
-                new Quote { Date = DateTime.Parse("2017-03-15"), Open = (decimal)224.44, High = (decimal)226.21, Low = (decimal)224.18, Close = (decimal)225.75, Volume = 101699816 },
-                new Quote { Date = DateTime.Parse("2017-03-14"), Open = (decimal)224.08, High = (decimal)224.13, Low = (decimal)223.14, Close = (decimal)223.81, Volume = 63382108 },
-                new Quote { Date = DateTime.Parse("2017-03-13"), Open = (decimal)224.49, High = (decimal)224.72, Low = (decimal)224.13, Close = (decimal)224.67, Volume = 60604724 },
-                new Quote { Date = DateTime.Parse("2017-03-10"), Open = (decimal)224.82, High = (decimal)224.87, Low = (decimal)223.52, Close = (decimal)224.56, Volume = 86785840 },
-                new Quote { Date = DateTime.Parse("2017-03-09"), Open = (decimal)223.62, High = (decimal)224.13, Low = (decimal)222.72, Close = (decimal)223.78, Volume = 95986360 },
-                new Quote { Date = DateTime.Parse("2017-03-08"), Open = (decimal)224.23, High = (decimal)224.51, Low = (decimal)223.34, Close = (decimal)223.49, Volume = 82739456 },
-                new Quote { Date = DateTime.Parse("2017-03-07"), Open = (decimal)224.25, High = (decimal)224.64, Low = (decimal)223.68, Close = (decimal)223.91, Volume = 68910464 },
-                new Quote { Date = DateTime.Parse("2017-03-06"), Open = (decimal)224.38, High = (decimal)224.97, Low = (decimal)223.92, Close = (decimal)224.58, Volume = 58630368 },
-                new Quote { Date = DateTime.Parse("2017-03-03"), Open = (decimal)225.01, High = (decimal)225.43, Low = (decimal)224.6, Close = (decimal)225.25, Volume = 86767480 },
-                new Quote { Date = DateTime.Parse("2017-03-02"), Open = (decimal)226.33, High = (decimal)226.34, Low = (decimal)225.05, Close = (decimal)225.11, Volume = 74353376 },
-                new Quote { Date = DateTime.Parse("2017-03-01"), Open = (decimal)225.22, High = (decimal)227.04, Low = (decimal)225.2, Close = (decimal)226.53, Volume = 157879712 },
-                new Quote { Date = DateTime.Parse("2017-02-28"), Open = (decimal)223.6, High = (decimal)223.86, Low = (decimal)222.98, Close = (decimal)223.41, Volume = 102631472 },
-                new Quote { Date = DateTime.Parse("2017-02-27"), Open = (decimal)223.57, High = (decimal)224.2, Low = (decimal)223.29, Close = (decimal)224.01, Volume = 59819992 },
-                new Quote { Date = DateTime.Parse("2017-02-24"), Open = (decimal)222.45, High = (decimal)223.71, Low = (decimal)222.41, Close = (decimal)223.66, Volume = 87198608 },
-                new Quote { Date = DateTime.Parse("2017-02-23"), Open = (decimal)223.79, High = (decimal)223.81, Low = (decimal)222.55, Close = (decimal)223.38, Volume = 78978816 },
-                new Quote { Date = DateTime.Parse("2017-02-22"), Open = (decimal)222.98, High = (decimal)223.47, Low = (decimal)222.8, Close = (decimal)223.23, Volume = 65747160 },
-                new Quote { Date = DateTime.Parse("2017-02-21"), Open = (decimal)222.51, High = (decimal)223.62, Low = (decimal)222.5, Close = (decimal)223.43, Volume = 94146880 },
-                new Quote { Date = DateTime.Parse("2017-02-17"), Open = (decimal)221.03, High = (decimal)222.1, Low = (decimal)221.01, Close = (decimal)222.1, Volume = 81718352 },
-                new Quote { Date = DateTime.Parse("2017-02-16"), Open = (decimal)221.98, High = (decimal)222.16, Low = (decimal)220.93, Close = (decimal)221.75, Volume = 89676304 },
-                new Quote { Date = DateTime.Parse("2017-02-15"), Open = (decimal)220.55, High = (decimal)222.15, Low = (decimal)220.5, Close = (decimal)221.94, Volume = 91860344 },
-                new Quote { Date = DateTime.Parse("2017-02-14"), Open = (decimal)219.71, High = (decimal)220.8, Low = (decimal)219.33, Close = (decimal)220.79, Volume = 75266840 },
-                new Quote { Date = DateTime.Parse("2017-02-13"), Open = (decimal)219.26, High = (decimal)220.19, Low = (decimal)219.23, Close = (decimal)219.91, Volume = 58408632 },
-                new Quote { Date = DateTime.Parse("2017-02-10"), Open = (decimal)218.24, High = (decimal)218.97, Low = (decimal)217.88, Close = (decimal)218.72, Volume = 69875952 },
-                new Quote { Date = DateTime.Parse("2017-02-09"), Open = (decimal)216.88, High = (decimal)218.19, Low = (decimal)216.84, Close = (decimal)217.86, Volume = 69811760 },
-                new Quote { Date = DateTime.Parse("2017-02-08"), Open = (decimal)215.98, High = (decimal)216.72, Low = (decimal)215.7, Close = (decimal)216.58, Volume = 54581376 },
-                new Quote { Date = DateTime.Parse("2017-02-07"), Open = (decimal)216.71, High = (decimal)216.97, Low = (decimal)216.09, Close = (decimal)216.29, Volume = 61318484 },
-                new Quote { Date = DateTime.Parse("2017-02-06"), Open = (decimal)216.23, High = (decimal)216.66, Low = (decimal)215.92, Close = (decimal)216.28, Volume = 61169192 },
-                new Quote { Date = DateTime.Parse("2017-02-03"), Open = (decimal)216.18, High = (decimal)216.87, Low = (decimal)215.84, Close = (decimal)216.67, Volume = 85273832 },
-                new Quote { Date = DateTime.Parse("2017-02-02"), Open = (decimal)214.65, High = (decimal)215.5, Low = (decimal)214.29, Close = (decimal)215.19, Volume = 73730552 },
-                new Quote { Date = DateTime.Parse("2017-02-01"), Open = (decimal)215.65, High = (decimal)215.96, Low = (decimal)214.4, Close = (decimal)215.05, Volume = 83743792 },
-                new Quote { Date = DateTime.Parse("2017-01-31"), Open = (decimal)214.44, High = (decimal)215.03, Low = (decimal)213.82, Close = (decimal)214.96, Volume = 80317680 },
-                new Quote { Date = DateTime.Parse("2017-01-30"), Open = (decimal)215.57, High = (decimal)215.59, Low = (decimal)213.9, Close = (decimal)214.98, Volume = 84399624 },
-                new Quote { Date = DateTime.Parse("2017-01-27"), Open = (decimal)216.75, High = (decimal)216.91, Low = (decimal)216.12, Close = (decimal)216.32, Volume = 63202528 },
-                new Quote { Date = DateTime.Parse("2017-01-26"), Open = (decimal)216.73, High = (decimal)217.02, Low = (decimal)216.36, Close = (decimal)216.66, Volume = 63477304 },
-                new Quote { Date = DateTime.Parse("2017-01-25"), Open = (decimal)216.07, High = (decimal)216.89, Low = (decimal)215.89, Close = (decimal)216.89, Volume = 89374928 },
-                new Quote { Date = DateTime.Parse("2017-01-24"), Open = (decimal)213.89, High = (decimal)215.48, Low = (decimal)213.77, Close = (decimal)215.03, Volume = 101142584 },
-                new Quote { Date = DateTime.Parse("2017-01-23"), Open = (decimal)213.85, High = (decimal)214.28, Low = (decimal)212.83, Close = (decimal)213.66, Volume = 79450624 },
-                new Quote { Date = DateTime.Parse("2017-01-20"), Open = (decimal)214.18, High = (decimal)214.75, Low = (decimal)213.49, Close = (decimal)214.21, Volume = 136721344 },
-                new Quote { Date = DateTime.Parse("2017-01-19"), Open = (decimal)214.31, High = (decimal)214.46, Low = (decimal)212.96, Close = (decimal)213.43, Volume = 70503512 },
-                new Quote { Date = DateTime.Parse("2017-01-18"), Open = (decimal)214.02, High = (decimal)214.27, Low = (decimal)213.42, Close = (decimal)214.22, Volume = 57997156 },
-                new Quote { Date = DateTime.Parse("2017-01-17"), Open = (decimal)213.81, High = (decimal)214.25, Low = (decimal)213.33, Close = (decimal)213.75, Volume = 64821664 },
-                new Quote { Date = DateTime.Parse("2017-01-13"), Open = (decimal)214.21, High = (decimal)214.84, Low = (decimal)214.17, Close = (decimal)214.51, Volume = 66385084 },
-                new Quote { Date = DateTime.Parse("2017-01-12"), Open = (decimal)213.99, High = (decimal)214.22, Low = (decimal)212.53, Close = (decimal)214.02, Volume = 76329760 },
-                new Quote { Date = DateTime.Parse("2017-01-11"), Open = (decimal)213.86, High = (decimal)214.55, Low = (decimal)213.13, Close = (decimal)214.55, Volume = 79014928 },
-                new Quote { Date = DateTime.Parse("2017-01-10"), Open = (decimal)213.97, High = (decimal)214.89, Low = (decimal)213.52, Close = (decimal)213.95, Volume = 67500792 },
-                new Quote { Date = DateTime.Parse("2017-01-09"), Open = (decimal)214.38, High = (decimal)214.53, Low = (decimal)213.91, Close = (decimal)213.95, Volume = 49684316 },
-                new Quote { Date = DateTime.Parse("2017-01-06"), Open = (decimal)214.02, High = (decimal)215.17, Low = (decimal)213.42, Close = (decimal)214.66, Volume = 75744152 },
-                new Quote { Date = DateTime.Parse("2017-01-05"), Open = (decimal)213.77, High = (decimal)214.06, Low = (decimal)213.02, Close = (decimal)213.89, Volume = 82961968 },
-                new Quote { Date = DateTime.Parse("2017-01-04"), Open = (decimal)213.16, High = (decimal)214.22, Low = (decimal)213.15, Close = (decimal)214.06, Volume = 83348752 },
-                new Quote { Date = DateTime.Parse("2017-01-03"), Open = (decimal)212.61, High = (decimal)213.35, Low = (decimal)211.52, Close = (decimal)212.8, Volume = 96708880 },
-                new Quote { Date = DateTime.Parse("2017-01-02"), Open = (decimal)212.61, High = (decimal)213.35, Low = (decimal)211.52, Close = (decimal)212.0, Volume = 76708880 },
-                new Quote { Date = DateTime.Parse("2017-01-01"), Open = (decimal)212.61, High = (decimal)213.35, Low = (decimal)211.52, Close = (decimal)211.6, Volume = 86708880 },
-            };
-
-            return h.OrderBy(x => x.Date);
-        }
-    }
-}
